Reader module tests, documentation and interface overhaul (#21)
* Add documentation for reader module

* Re-export Reader in lib.rs

* Add tests for Reader and read header at creation

Implementing a lazy reading of the header causes a couple problems in
the implementation (e.g. infinite loops without a couple state flags)
and the writer schema to be an Option until the first value is read
or read_header is called. For this reason, header is read at creation
time and now both new and with_codec do return a Result.

Writing tests, a second bug that could cause many errors to be returned
by the iterator instead of only one emerged. To fix it, a flag to signal
that the reader entered in a state of error has been added, so the
iteration will stop automatically at the next step.

* Add from_avro_datum function and more tests for reader

from_avro_datum has been added for parity with to_avro_datum