Fix variable-length encoding for big i64 numbers
`util::encode_variable(i64)` took a i64 but used both `>>` (which does a
arithmetic shift for i64) and `>=`. Those operations don't work on
numbers/bytes which are still zig-zag-encoded because a large zig-zag
encoded number may have a leading `1`. This breaks `>=` as the number is
assumed to be negative and also the right-shift will keep introducing
new `1`s which leads to an infinite loop.

Replacing the argument (and the callsites) with a u64 fixes those
issues.

To make sure this does not regress in the future, tests for `i64::MAX`
and `i64::MIN` have been added.