mtrlz/sql: handle aggregates in WHERE/HAVING
Aggregates are explicitly banned from the WHERE clause and explicitly
allowed in the HAVING clause. The old version of the SQL parser was
somehow misparsing aggregates in either of these locations, and so was
hiding the fact that our planner couldn't handle aggregates in either
cases. Teach the planner to handle aggregates in either position--i.e.,
by rejecting aggregates in the WHERE clause and properly extracting them
in the HAVING clause.