mtrlz: unify INSERT code paths
Adjust the dataflow module's insert command to take an arbitrary
dataflow as its source, rather than a constant list of values. This
unlocks support for complex insert statements, like

    INSERT INTO t SELECT ...

where the SELECT ... query is any arbitrary query that we are currently
capable of executing.

This requires some timestamp shenanigans in order to perform the query
at time t and perform the insert at time t + 1. I plan to document this
scheme further soon.

The other major change is that the planner is no longer aware of how
many rows an INSERT statement will generate, and so cannot immediately
report this answer to the client. The mechanism for sending PEEK results
has been generalized so that it can also be used for reporting INSERT
counts to the client.