testdrive: don't reuse kafka topic names
Reusing Kafka topics is fraught with peril. (See the enclosed removed
comments for details; basically, Kafka deletion is asynchronous, and can
impact a topic created with the same name in the future.)

Instead, randomly generate a name for each testdrive invocation, like
testdrive-TOPICNAME-123456789. We still clean up after ourselves by
deleting any topics that start with testdrive-TOPICNAME on the next
testdrive invocation, but those topics are abandoned at that point and
so can't impact the current testdrive run.