mtrlz/sql: unify coalesce and cast code paths
There are two code paths in the SQL planner for converting values
between different types. The first code path (the "cast" code path) is
used by explicit CAST operators and the implicit cast that occurs when
inserting a datum of a given type into a column of a given type. The
second code path (the "coalesce" code path) is used when performing an
operation that implicitly coalesces types, like coalesce(a, b, c), or a
CASE statement with multiple branches, where the arguments need not be
of exactly the same type if they can be automatically converted into a
mutually agreeable type.

In what appears to be a historical accident, the two code paths had
completely independent rules for what types could be converted into
other types. The cast rules were more complete, so teach the coalesce
code path to simply call into the cast code path.