ci: speed up
Speed up CI by breaking the pipeline into multiple steps. The first
step, which builds Docker images for the various test binaries, is the
only blocking step. The other steps simply pull down the particular
Docker images they need and can run in parallel in one another.

This introduces quite a bit of complexity too, but it's complexity we'd
need to introduce down the road, anyway, in order to shard jobs.