.github: configure merge-when-green
Configure merge-when-green so that it knows when status checks have
passed. The idea is that you attach the "merge when green" label to a PR
that you want to be automatically merged when tests pass. Useful when
your PR has been LGTM'd, but you had to push fixes for a few nits and
don't want to sit around staring at GitHub waiting for the build to
pass.