dataflow: centralize exfiltration logic
PEEK and TAIL share a lot of logic about sending results back to the
client, but that logic was previously not easily shared. Centralize that
logic in a new exfiltrate module.

Also take the opportunity to fold DataflowCommand::Tail into
DataflowCommand::CreateDataflow, now that the connection UUID is
available to the SQL layer and not attached via a sidechannel
CommandMeta struct.