Add interval addition/subtraction for Dates and Timestamps
As soon as you interact with interval arithmatic you have to type promote up to
a Timestamp, no matter what your input type is, because Intervals can contain
times or monthlike spans.

This takes the Decimal-like approach and does a bunch of logic in
`plan_binary_op`, instead of maybe doing some sort of type coalescing in
`try_coalesce_types`, because that function is used for genuine coalescing, not
just "find the output type".

Closes #181