kafka: Always poll within the required interval
Kafka requires a that user threads poll the queue in _addition_ to a separate
heartbeat thread that keeps the cluster aware of the client. The idea [seems to
be][1] that if a user thread is too overwhelmed to poll occassionally then kafka
should rebalance things.

Different versions of kafka have the default at different values, but the
current default is 5 minutes, so we set it to that.

Then, we poll once per minute just to be really sure that we are never evicted.

[1]: https://stackoverflow.com/a/43992308/25616