comm: interleave messages from multiple MPSC senders
Fix a bug in the implementation of MPSC senders in which messages from
producers were presented in a total order, i.e., messages from the
second producer would only be presented after the first producer was
closed. Now messages from multiple producers are properly interleaved.

This required the addition of a new `SelectFlatten` stream combinator,
since nothing like this exists in the futures library itself.