comm: use well-typed addresses, rather than strings
This also fixes the defect where a MPSC transmitter could be connected
with the wrong type. E.g., after the last commit, it was possible to say

    let switchboard = Switchboard::new::<TcpStream>();
    let (tx, rx) = switchboard.mpsc();
    tx.connect::<UnixStream>();

which would then crash loudly because a TCP socket address is not a
valid Unix socket address. The transmitter now carries with it the type
information about what type of connection it is supposed to establish,
to avoid such mistakes.

The cost is a bit of flexibility, as it is no longer possible for
someone else to show up with an implementation of `protocol::Connection`
and have the comm library seamlessly work. That's fine, though, since we
control the comm library and can just add a first-party trait
implementation and `protocol::Addr` variant.