sql: support ordinal column refs in ORDER BY
SQL weirdly supports using ordinal column references in ORDER BY, like

    SELECT a, b FROM x ORDER BY 2, 1

where the 1 and 2 refer to the columns named a and b, respectively. The
sqllogictests make extensive use of this notation. Adding support for
this notation is quite easy and should yield a massive increase in our
sqllogictest successes, now that most of Jamie's subquery work is in.