expr,sql: map via joins
Frank had the insight that mapping via joins with a constant relation
expression allows our optimizer to do more work, including eliding the
map/join entirely if predicates can be pushed down. Use this pattern in
another place in the decorrelation code, to save on a redundant
map/filter, and also leave some comments behind in both places it's
used.