expr: standardize rendering of constant nodes
While debugging some query plans, it's been pretty confusing that both
an empty constant and a constant with an empty row were rendered as (),
and that a constant with a single row rendered as (1) while a constant
with two rows rendered as ((1), (2)). While nice in isolation, I think
the brevity is actually impeding understanding here.

This commit adopts a new standard output format that looks like
array/slice notation in many languages.

A constant with no rows:                []
A constant with an empty row:           [[]]
A constant with one row of one column:  [[1]]
A constant with two rows of one column: [[1], [2]]

While a bit more verbose, this will be quite helpful to me while
debugging query plans!