Build docs and dependencies
This makes autolinks to our dependencies work. The bug with multiple
versions of the same crate isn't hugely problematic; `cargo doc` will
just randomly pick one of the docs to complete. This means the
occasional link will be broken, but that's better than all of the links
to our dependencies being broken.