expr: remove RelationExpr::Distinct
RelationExpr::Distinct can be trivially implemented in terms of
RelationExpr::Reduce by specifying no aggregation functions. Do that, to
cut down on the number of nodes we need to reason about.

We may want to hold off on implementing this until #573 is done, as we
currently have some optimizations that apply easily to distinct but
require a bit more elbow grease to apply to reduce.