expr: use sequential ID allocation when planning
This makes it possible to write static assertions about the expected
query plan. The previous approach of using UUIDs meant that the IDs
changed on every replan.

Note that the additional ID cleanup transformation in #594 is still very
much warranted, as expr transformations will leave holes in the ID
allocation that can be cleaned up with that transformation.

Now that all query plans are deterministic, fill in some more expected
chbench plans.

Fix #608.