expr: add compact pretty printing of lists
This is kind of a hack, and doesn't always get things exactly right, but
it does a decent job and is much better than printing long lists of
outputs
one
per
line
which
is
really
painful
when
there
are
forty
plus
outputs. See the doc comment within for details on the approach.