expr,sql: use distinct_by in more places
Now that we have an explicit `RelationExpr::distinct_by`, the old idiom
of achieving distinct by via `RelationExpr::project` followed by
`RelationExpr::distinct` is deprecated, as it introduces an unnecessary
project node. Update a few call sites to use the new idiom.

One day we'll want to do this via a transformation, too, but it's useful
to just construct the desired plan directly, when it's easily within our
control to do so.