[wip] add a script to cross-compile materialize to linux
The intended eventual flow is to run build.sh from the materialize directory to generate
a cache of the the dependencies:

    bash misc/docker/cross/build.sh

and then to run the rebuild script to create the binary file in
target/linux/release/materialized:

    bash misc/docker/cross/rebuild.sh

or just use `materialize/local-build` as a docker-image or FROM target.

Then... git checkout Cargo.toml to get rid of the hacks in the script.