bin/ci-image: clean up docker image naming
The built images should not use the "ci-" prefix, because we're not
actually in CI; the ci- prefix is just used to stage the actual images
before they're published. Also teach the run subcommand to use the
`local` tag, just to increase the probability that the right image gets
selected. Previously it was possible to accidentally run an image from
the registry, rather than one that was locally built.