sqllogictest: speed up postgres integration
It was taking multiple minutes to insert 10000 rows into Postgres. The
issue was that we'd diff the whole table after every
INSERT/UPDATE/DELETE statement. We can be quite a bit smarter than that,
by asking Postgres to tell us only about the rows that were
inserted/updated/deleted via the `RETURNING *` Postgres extension.  Note
that updates are a tad more involved, as we need to run a separate query
to determine the pre-update values.

This makes inserting 10000 rows take a few seconds, which is far more
reasonable.