comm: reuse connections for MPSC senders
Since most channels are used by worker nodes to send peek results to the
coordinator, it's a waste to open a new TCP stream for every peek. Teach
the MPSC senders to use a thread-local connection pool instead.

This required some reworking of the handshake protocol so that
connections can be disassociated with their current receiver and
connected to a new one.

Fix #644.