expr,sql: learn to derive a ScalarExpr's types
Previously, the SQL planner would keep track of a scalar expression's
types as it built up the scalar expression, throwing away type
information about the leaves of the expression as it worked its way
towards the root. This is problematic when a ScalarExpr later needs to
be split into pieces, as the type information about the inside of an
expression was lost.

Add a new ScalarExpr::typ method that can derive a scalar expression's
type on the fly. This requires that ScalarExpr::Literal contain type
information, as the full type of decimal and null literals cannot be
inferred.

Note that ScalarExpr::typ is added to both ::sql::expr::ScalarExpr and
::expr::ScalarExpr. The ::sql::expr version requires both the outer and
inner relation expressions that the scalar expression will be evaluated
upon (due to correlation), while the ::expr version requries just the
single relation expression (thanks to decorrelation).

Then, teach the planner to use ScalarExpr::typ whenever scalar
expression type information is required, rather than passing it upwards
along with the expression. Functions that planned scalar expressions
previously returned (ScalarExpr, ColumnType), and now return
ScalarExpr, as the column type can be computed (via typ) whenever
required.

Finally, remove the ColumnTypes from RelationExpr::Reduce and
RelationExpr::Map, as they can be derived from the contained
ScalarExprs/AggregateExprs as necessary. This prevents them from going
stale as optimizations rewrite the tree.

Addresses half of #716.