coord,dataflow: only produce future updates in TAIL
This supersedes #427. The feedback in that PR was substantially easier
to address after the coordinator refactor. TAIL now only produces
records *after* the TAIL command begins, so that TAILing a large
relation does not require outputting a snapshot of that large relation
first. The biggest difference is that the notion of current time is
derived from the upper frontier of the tailed view, rather than using
the current system time.

Fix #21.