pgwire: fix DescribePortal handling
DescribePortal handling was previously broken; it was reporting
parameter statuses, but only DescribeStatement is meant to report
parameter statuses. And then I accidentally removed it entirely in #784.

Restore it, and this time with the proper behavior of not reporting
parameter statuses, but only row descriptions.

This is required for PgJDBC.