coord: better leverage type system in API
The coordinator's API previously assumed that all `coord::Command`s
needed a connection ID, response transmitter, and SQL session, and that
a `coord::Response` would always be sent over the response transmitter
after processing was complete.

This assumption had a number of exceptions. CancelRequests, for example,
need only a connnection ID and produce no response; parse requests need
only respond with failure or success, and have no need to return the
kind of SQL statement they parsed to.

Rework the coordinator API so that each command declares exactly what
fields it needs and what fields it returns. This results in a bit of
duplication, especially between Query and Execute commands, but allows
for much greater flexibility, and a more descriptive API.

Also take the opportunity to introduce a new Shutdown command, which
shuts down the coordinator thread. This will be used in the forthcoming
"symbiosis" mode, where the SLT runner will drive a coordinator via the
same `coord::transient::serve` interface that materialized uses, rather
than assembling a bespoke dataflow server out of its components.