All names are QualNames
This add support for dotted and quoted identifiers. In particular, we need to have a way
to make the following equality hold, when entered at a terminal:

    oNe."two"."tHRee" == one.two."tHRee"

This introduces a `repr::QualName` type (in `repr/qualname.rs`) that stores how data was
entered, but does comparison based exclusively on normalized values. There are lots of
things that could be done to make this perhaps more ergonomic and/or efficient.

In terms of ergonomics I basically just did the smallest things that I could to get
QualNames working, ideally we will delete a bunch of the inherent methods over time.

In terms of efficiency these are probably significantly less efficient than the old
string implementation, but now that the actual representation of the names are hidden
there are some pretty obvious things we could do to improve things without changing the
external API.

Fixes #685