Changes frontier reporting to event driven.
This PR changes the way workers report their frontier information back to the coordinator, and how the coordinator uses this information. Previously, all workers would report the upper frontiers of all traces they manage. The coordinator would select the frontier of worker zero, and use that as the truth. This information drives the timestamps at which we peek, so it isn't correctness critical, but it can stall peeks a bit if it is unintentionally in advance of some other worker.

The new logic changes all of the communication to be event-driven, with workers communicating changes back to the coordinator. The coordinator now uses the lower bound of frontiers from all workers, which has the potential to unblock some peeks that might have otherwise been blocked.