Enable a subset of the cockroach/case_sensitive_names tests
Our names come back with quotes around them, so that changed, not sure if we want to have
a different display method for returning to clients,

I had to use some other function rather than `length` for the case-ignoring-function
test, so I used ascii, which means that we now convert characters into strings in SLTs.

I'm actually not sure what the semantics of `ASCII` are supposed to be, should that be
returning a string to clients?