materialized: various integration test adjustments (#1008)
* materialized: assert number of rows in prepared test

The previous test would erroneously succeed if the query incorrectly
returned zero rows.

* materialized: let kernel choose free port

Rather than trying to pick a random port for materialized to listen on
in tests, let the kernel choose a free port by specifying port 0, and
communicate the chosen port back to the test.

* materialized: don't allow configuring the version

The crate version is a property of the crate itself, so it was kind of
weird to allow the user of the crate to pass in the version. Invert that
flow, so that the crate hardcodes its own version, and downstream
dependencies can query its version as necessary.