Respect `since` bounds.
This PR modifies timestamp selection to respect the `since` bounds used by compacted sources. Rather than simply picking the largest time that can be immediately answered, we also restrict our attention to times greater than all `since` bounds for each involved trace. Ideally this has no observable outcome, but it has the potential to change silent failures into errors (that no time can be found, instead of querying at a time for which we may produce an incorrect answer).

There is currently no flow for error reporting when we fail to find a timestamp, and that should probably be added before we land this. I'll need to learn a bit more about how to add those, though, or get some help.