Make runv stick out more in logs
The `run>` word just looks a lot like the text that it is often interspersed with. I
wanted something that would stick out but not be too distracting, I'm happy to change
this to something less cute or more functional if anyone has ideas (I read this prompt as
"launch shell command") but the plain white text was annoyingly hard to spot when just
glancing.