grafana: Make peek durations/latencies more useful
This adds a grafana heatmap visualization for both client and server peek durations, and
makes the server peek durations... readable.

This uses the grafana histogram panel which you should read the first two sections of
[the docs] for because it can be pretty confusing.

This also makes the client side and server side views a bit different, because grafana
cannot automatically subtract the values between two histogram buckets, and the
underlying data format is annoying, but this allows us to have a slightly more accurate
view on the server.

Fixes #896
Hopefully it helps a bit with #1074

[the docs]: https://grafana.com/docs/features/panels/heatmap/