catalog,coord: install logging views before loading stored entries
Loading a catalog would crash if the catalog contained entries that
depended on logging views, because the logging views were loaded after
the persisted catalog entries. Switch the order to fix this.

Also add a test case for this, and for a somewhat related bug with the
bootstrap SQL feature, in which passing `--bootstrap-sql` when a cluster
was already bootstrapped would fail.

Fix #1115.