Improve names of logged collections.
Our collections had names that described their source of information, rather than names that hinted at what they contained. This PR attempts to improve, if not fully solve, this situation with a `mz_` prefix and somewhat more relevant names. There is still a rationalization required to bring together the contents of all of this information, but that is deferred for now.