sql: support naive variance and stddev computations
Add support for computing the stddev of float and decimal data using the
"textbook" algorithm. This algorithm is known to accumulate problematic
amounts of error, but it is much more straightforward to implement in
differential dataflow than any of the numerically-stable algorithms. In
fact, it requires no changes to the dataflow layer at all, as the SQL
function can simply be decomposed into the sum, count, and sqrt
functions.

It's also worth noting that PostgreSQL got away with using this
algorithm until version 12, released last year.

Implementing a more accurate variance/stddev algorithm is tracked in
issue #1240.