peeker: separate recording of errors and timings
It isn't super interesting how long our errors take -- they're essentially always either
query-does-not-exist or cannot-connect, which are falsely fast.

So this eliminates those measurements, and records errors instead.

In order to get access to that I needed to bump to the recent git for prometheus, which
nicely eliminates a few dependencies/different versions of dependencies.