xcompile: tell cmake to cross compile
CMake needs to be explicitly told that it is cross compiling by setting
the CMAKE_SYSTEM_NAME define.

We were actually doing this entirely wrong before; we were installing
the macOS SDK headers so that the macOS version of endian.h would be
picked up by CMake. This only happened to work because we were
cross-compiling from x64 macOS to x64 Linux. If we properly tell CMake
that we're cross-compiling, it won't accidentally try to look for
headers in the host's /usr/include. A recent change in rdkafka made it
so that things failed hard, rather than worked with subtle bugs.

This also has the side effect of fixing compilation on macOS 10.15,
which no longer provides a way to install headers into /usr/include
anyway.