dataflow: properly handle source lifetimes
The file source really wants the promise that it will never be scheduled
again if it indicates that it is done, i.e., by returning
SourceStatus::Done. Timely is unwilling to make this promise, but it's
easy to teach the `util::source` helper to make this promise.

Implementation details are in the comment within.