pgrepr: new crate to handle pgwire binary encoding
Extract a new crate, pgrepr, that will handle the pgwire binary
encoding. The rationale for this crate is laid out in #1380, but the
summary is that we have several crates that want to use the binary
encoding/decoding logic but cannot or should not depend on the pgwire
crate.

This is bare minimum code movement for now. Forthcoming commits will
flesh out the crate.

Touches #1380.