pgrepr: take responsibility for decoding values
Move responsibility for decoding datums from pgwire to pgrepr. pgrepr
delegates the binary decoding to postgres-types where possible, just
like it does for encoding.

This required introducing a new pgrepr::Type object that is a direct
mapping to a PostgreSQL type. It differs from repr::ScalarType in only
one meaningful way, which is that pgrepr::type::Numeric does not store
the scale of the numeric, while repr::ScalarType::Decimal does. This
mirrors how pgrepr::Value::Numeric stores the scale, while
repr::Datum::Decimal does not.

A nice side effect of pgrepr::Type is that it makes the conversions
explicit. It is always possible to go from a repr::ScalarType to a
pgrepr::Type, because the latter stores *less* information. But going
from a pgrepr::Type to a repr::ScalarType requires the actual value to
be on hand, so that the decimal scale can be filled in if applicable.

pgrepr::Type delegates to postgres-types where it can, and completely
subsumes the old pgwire::types module.

Despite the number of todos in this patch, the only regression from
previously-supported behavior is decoding of textual bind parameters,
which aren't tested anyway.