catalog: take responsibility for QualNames
Qualified names are fundamentally a SQL/catalog concept; the dataflow
layer is intended to be entirely agnostic to names. Move QualNames out
of repr, which is intended to be SQL-agnostic, and into catalog.

This is the first step towards supporting proper databases and schemas
(#1126). It is also the second-to-last step for breaking repr's
dependence on the sql-parser crate (#1006).