Allow reduct to use ScalarExpr as keys.
This PR generalizes the keys of RelationExpr::Reduce to be a sequence of ScalarExpr, rather than just a sequence of column indicies. This allows the Reduce operator to handle key production on its own, without requiring a Map operator to produce keys for it. This is mostly a superficial change at the moment, as the only way to construct Reduce stages is using column identifiers, because that is what planning expects. However, we could generalize that in the future, and we can currently go ahead and transform the query plans to absorb Map stages into Reduce stages.

On ergonomic regression, which I took a peek at fixing but didn't understand, is that we used to tighten sequences of column identifier to be the first and last elements with elipses between. This was fine when we were creating a string, but if we need to use the document builder to do this it starts to get gross and I'm not trained on how to do that. So, we have the opportunity to change things like #0, #1, #2, #3 into #0 .. #3.