Support handling errors in the RowPacker push_*_with methods
Add a new method `try_push_dict_with` that supports throwing errors. Since it is not
possible to know from outside the closure whether the dict is valid -- if the failure
happened in between writing a key and a value, for example -- it consumes the rowpacker
for errors.

Similarly, this adds a `try_push_list_with` method that supports returning errors, and
guarantees that the list is finished afterwards. Since it's impossible for the individual
packer `write` commands to leave the list in an inconsistent state, we *could* sort of
YOLO clients and allow them to continue using the rowpacker if they think it's fine,
unfortunately that makes recursive calls with `try_push_dict_with` impossible, so we need
to add this method and give it the same API.

This will allow a bit of a refactoring in the next commit.