Make RelationExpr::TopK more robust.
This PR changes the behavior of TopK to do hierarchical reduction, repeatedly grouping modestly sized groups and applying the ordering and limit. This avoids massive reductions that require significant work on individual updates, at the cost of a more complex dataflow, and potentially more memory maintained for those arrangements that do not actually reduce the data. This is expected to be very beneficial in the case of "top k" sorts of queries, and less helpful for subqueries or other settings where large groups are not formed.

The motivation for the PR is that robustness is the best default behavior, and optimizing constants is something that can be done by experts in an opt-in manner. We can negotiate that, but this does absolutely remove a massively defective default experience for order by / limit queries.