dataflow: don't miss file updates on macOS
notify::PollWatcher can miss updates because it uses an mtime resolution
of 1s. If a file changes twice within 1s, we could get woken up for the
first update but not the second. Rather than trying to use the
PollWatcher, just wake ourselves up every 100ms and check if the file
has new data; this actually *saves* work, and doesn't have the
aforementioned bug.