materialized: restore intent of test_file_sources
In the switch to non-materialized sources, this test was refactored to
repeatedly create new views, which would cause the file to be repeatedly
re-read, in the case of dynamic files. This was not the intent of the
test--the test is meant to ensure that tailed files update without the
source being recreated--and the new version was masking the bug fixed in
the previous comment.

Restore the original intent of the test by materializing the source via
one view that lasts for the duration of the test.