testdrive: derive descriptor data from compiled rust protos
For protobuf sources, Materialize needs the FileDescriptorSet that
describes the protobuf schema. Testdrive was currently producing this
data twice; once in the build script, when compiling the Rust protobufs,
and again at runtime. Just expose the FileDescriptorSet that's already
embedded within testdrive, which allows us to eliminate a *lot* of
complexity. Notably, the data directory, which didn't compose well with
stdin scripts, is no longer necessary.