testdrive: remove superflous error variants
ErrCtx already exists with exactly the same behavior in the same module.
It's called ResultCtx.

Error::Message is a less-general version of Error::General. The latter
is intentionally not constructible from a string, in the hopes that
folks will provide more detailed error messages rather than using
whatever string happens to be convenient.