sql: don't confuse inline and file protobuf schemas
The statement

    CREATE SOURCE s FROM 'kafka://...' USING SCHEMA 'foo'

would previously interpret 'foo' as an inline schema whenever 'foo' was
valid base64, and a path to a file otherwise. This is a bad assumption,
as something like 'filename' would be incorrectly interpreted as a
protobuf descriptor with the contents 'filename', rather than the path
to a file named 'filename'.

The fix in this commit is to introduce a separate statement variant for
reading from disk, i.e.,

    CREATE SOURCE s FROM 'kafka://...' USING SCHEMA FILE 'foo'

so that the first form unambiguously refers to inline schemas.

This commit also changes the expected encoding of the inline schema from
base64 to the PostgreSQL bytea input format, for consistency with the
existing support for representing arbitrary bytes as string literals.

The concept of a schema file is fairly general, so this commit
allows Avro schemas to be read from files as well.

Finally this commit adjusts the protobuf connector to store the raw
descriptors, rather than storing the path to the descriptor file. This
is important because we can't trust that the descriptor file will
continue to exist if we crash and reboot. Storing it in the connector
means it will be persisted to the catalog at plan time, and therefore is
guaranteed to continue to be accessible after a reboot.