*: remove bootstrap sql
There is no user need for a bootstrap SQL file now that we have
persistent metadata. We were using it internally to install a dozen or
so useful views over the logging sources, but it's just as easy to
hardcode those views inside Materialize, so that they are immediately
accessible to all users.

Fix #1528.
Fix #1601.