coord: refresh type for view descriptions.
We now refresh the type when we optimize the view, which folds in any new information
about unique keys, non-null columns, and things like this. This was previously blocked
on unintended removal of type information relation to nulls, and this change may now
exercise these changes, and possibly other defects.