doc-release: Update release process to use tags and single release branch
After the recent meeting, we have a possibly simpler release process, it certainly has
fewer branches.

Since the point of the `release` branch is to just make it easier to find the most-recent
release I did change the process slightly from the meeting: instead of merging the
`release` branch with master occasionally, I have it documented as just being reset to
the next release, I know that this will cause fewer merge conflicts (since it never
involves any merges) but it does mean that not all releases will be in the release
branch, which may be a problem.