sql: allow planning with an internal session
When catalog persistence is implemented in terms of SQL, we don't want
the persisted SQL queries to be reliant on session state. Put another
way, on boot up, there is no user session to use to plan the catalog's
SQL queries.

Introduce a PlanSession trait that is a restriction of a sql::Session to
just the parameters that control the planning of a SQL statement. Then
introduce a new InternalSession implementation that hardcodes sensible
defaults for the plan-time session parameters. So far we just unset the
session database and use an empty search path, which has the desired
effect of requiring all table names to be fully qualified.