doc-release: Update to use release-X.Y branches with no LTS
In particular, this doesn't ever merge from master, once we are confident in master we
just hard-reset to the commit that we would like to use. This maintains the
mainline-first development philosophy with the ability to stage changes.

Also add some hand-drawn images to give a general idea of what things look like, although
the branches and tags are all abreviated.