http-demo: speed up and simplify startup by using a cached image
This cuts an additional minute of time off initial startup and adds a `materialize/http-demo`
image that contains the dependencies for both the client and server.

In addition to making it twice as fast for me, this reduces the number of systems that we
depend on (now it's just a bunch of docker pulls, not docker/apt/pip) so it is hopefully
more robust.