catalog: become the source of truth for indexes
The optimizer needs to know about what indexes are available in order to
produce good query plans. The coordinator was previously computing this
information, but it's just as easy for the catalog to do it.

Putting the catalog in charge of keeping this metadata means that
restoring a catalog on boot can properly use this index information as
it parses, plans, and re-optimizes the SQL it stashed away. This wasn't
previously necessary because we stored pre-optimized plans on disk; this
is actually a nice improvement, because upgrading Materialize will also
re-optimize any stored plans, hopefully with an improved optimizer!