peeker/chbench: support a config file for multiple views
This additionally adjusts the load-test commands to use peeker instead of chbenchmark for
the peek threads.

Using peeker instead of chbenchmark to run queries gives us per-query client-side
latencies in the grafana dashboard, which is pretty nice.

Right now every query is executed in a tight loop in its own thread, but we could change
that to reduce or adjust load.