catalog: persist all ID allocations
Global IDs could previously get reused after a crash, because upon
reboot the ID allocator was reset to one greater than the maximum ID
present in the catalog. This was once safe, because any state that
referred to those reused IDs would have evaporated during the crash, but
causes trouble for the timestamp persistence code, which might persist
messages referring to those recycled IDs.

Instead, have every ID allocation durably record the new state of the ID
allocator, so that we can guarantee IDs are not reused even after a
reboot. This has potential performance implications, as there is now an
fsync on the path for slow-path peeks (not fast-path peeks, though!),
but there are bigger fish to fry with slow-path peeks anyway.