catalog,coord: share sqlite connection
Opening two SQLite connections leads to conflicts between the main
coordinator thread and the timestamper thread. According to the SQLite
documentation, the correct way to handle this is by sharing one
connection via a mutex across the two threads.

This should fix the "database is locked" error we're seeing with v0.1.