materialized: log to file by default
Logging to stderr is noisy, and should not be the default for folks
manually running the materialized binary. Make the default logging to
DATADIR/materialized.log, but allow overriding with
--log-file=<stderr|PATH>. Our Docker images are changed to set the log
file to stderr by default, to comport with Docker logging conventions.

Note that this commit also switches us to use the tracing crate
ecosystem, as env_logger does not support logging to files.