dc.sh: make load test more repeatable and easier
Most importantly from an external perspective this adds a `clean-load-test` top-level
command that runs `nuke` and then `load-test` with all dependencies. This is what we
should always run when qualifying a release.

In support of that, this adds `dc.sh load-test --up`, which runs the things that were
previously in `dc.sh up :load:`.

Because IMO it was confusing that we had different categories of commands that did
similar but not identical things. Now load tests are always run exclusively by
`load-test` commands.

There are several *bugfixes* in this commit:

* we were running `peeker` as a service as part of `up :load:`, which was reporting its
  metrics instead of the metrics we were explicitly requesting as part of the `load_test`
  function
* docker-compose run *by itself* does not expose services internally via dns, meaning
  that when running the load test, you'd only get the metrics from the peeker that was
  run as part of the `up :load:` command.
* running `docker-compose run` would cause `docker-compose` to do something to all
  `depends_on` services, *possibly* blowing away data, causing chbench to give up.

  I removed the depends_on for the two load-test services (chbench and peeker) because we
  really don't want docker-compose futzing with materialized/connect/etc just because
  we're trying to run a command
* chbench was coming up and *staying up* for between 6 and 13 seconds for me before
  crashing due to that "maybe data is missing" problem, so I added a wait-for-awhile
  function

I also added a wait-for-arbitrary-command function, because chbench was crashing because
mysql was getting.. restarted, or something.