Reduce the application of simplification (#2204)
This PR restricts the application of the simplification transformation (which unpacks equality predicates on complex expressions, so that they can be used as join keys). It now ensures that it is only applied when each of the two expressions in an equality depend on exactly one input collection and the two are not the same input collection (otherwise, the constraint can, and probably should, but pushed down into that collection).

Changes are much smaller with whitespace ignored. Lots of code got one additional indentation level.

fixes #761