expr: allow lazy evaluation of function arguments
AND, OR, and coalesce all want to evaluate their arguments lazily rather
than eagerly. This has semantic implications when evaluation can error.
For example

    coalesce(1, 1 / 0)

does not result in an error, because `1 / 0` is not evaluated,

    coalesce(NULL, 1 / 0)

does result in an error, because the `1 / 0` will be evaluated.

This patch installs the necessary plumbing to allow functions to opt-in
to lazy evaluation, and then teaches AND, OR, and coalesce to do so.