sql: accept SET TIME ZONE UTC
This is a small step towards allowing clients to configure their time
zone. Most clients want to use UTC, and some will issue a SET TIME ZONE
UTC command regardless of whether the connection is already using UTC.
Just accept that declaration, so we work with those clients, but error
if clients try to select a different time zone.

Fix #2393.