interchange: rework avro encoding
Rework how Avro encoding works so that it is no longer fallible. Rather
than generating a schema and encoder separately, have the encoder
tightly manage the generation of the schema; then encoding can rely on
these tightly-managed properties of the generated schema.

As a side effect, this diff adds support for encoding all known
Materialize data types as Avro.

This could use quite a few more tests, but those will have to wait for a
future commit that re-enables testdrive testing of Avro sinks.

Fix #1517.