coord,sql: tail IDs, not catalog entries
Returning a CatalogEntry in a sql::Plan is strange, since it could get
out of sync with the entry in the actual catalog. Since only the ID is
needed, I think it's cleaner and cheaper to pass around the GlobalId
instead.

Also make a match statement exhaustive while I'm in here, just to
future-proof the code. If we ever get a new type of tailable object,
it'll be harder to forget about it.