Make Avro Writer own Schema
Previously, avro::Writer held a reference to Schema and was bound by the borrowed
Schema's lifetime. We were not able to use Writer within a closure for dataflow
operators because it required that the borrowed Schema have 'static lifetime.

Instead, we changed Writer to directly own the Schema it uses, which has a
marginal cost on memory usage but leads to a much more broadly usable API