testdrive: rework kafka ingestion transcoding
Testdrive now supports ingesting records with both keys and values into
Kafka. Since `kafka-ingest` expects one line per Kafka record, that line
needs to contain both the key and value. This is easy enough when the
key or value format is Avro/Protobuf, as testdrive expects Avro/Protobuf
input in JSON, and JSON is self delimiting. A line like

    {"a": 1}  null

is easily separated into the key JSON ({"a": 1) and the value JSON
(null).

The bytes format is challenging, though, as the nature of an
unstructured format like bytes means that there is no one universal
separator that can be used to separate the key and value. Presently
testdrive requires that bytes input be a JSON-formatted string, but this
is unnecessarily limiting, as not all bytes input is representable as a
JSON string.

Instead, adopt the approach that kafka-console-producer uses of allowing
a configurable separator between key and value. By default the separator
is a `:` character, as in

    key: value

but it can be configured via the `key-terminator` option:

    $ kafka-ingest key-format=bytes key-terminator=$
    key$ value

Also rename the encoder to "transcoder" to capture the confusion that
Andi and I had about whether it was a decoder or encoder: it's actually
both.