testdrive: harden sink verification
We guarantee order for sinks now, so we should test that. Turns out
the fix in #2551 was unnecessary.

Also includes a few assorted cleanups that I noticed, most notably
timing out in 15 seconds if we're unable to read the expected number of
messages from Kafka, rather than waiting around forever.