Use timestamp zero for tails of unmaterialized sources/views (#2520)
This PR changes the behavior on TAIL requests for collections that are not materialized. Previously, we would error as we could not determine a time to tail from. Now, we use zero as the time. This PR also changes the behavior of TAIL to include updates at the current upper frontier of materialized collections, to align with the updates that are reflected by a SELECT statement at the same logical time.

fixes #2519