expr,sql: replace EvalEnv with PlanContext
EvalEnv represented plan-local (or dataflow-local) state, like the
logical time chosen for the query's execution, or the wall time at which
the query was planned. It was intended to one day store information
from the session that planned the query, like the session timezone,
which impacts how e.g. timestamptzs are parsed and formatted.

Unfortunately the approach made cross-dataflow optimization hard. Two
plans will almost always have different wall times, and in the future
might have different timezones... but most plans don't actually
reference the wall time (or the timezone, in the future), and so it's
perfectly safe to share dataflow fragments across these dataflows, even
though their EvalEnvs differ.

Instead, we can inline the EvalEnv into the AST.

So

    ScalarExpr::CallNullary(NullaryFunc::Now)

becomes:

    ScalarExpr::Literal(<inlined now>).

If we had timezones,

    ScalarExpr::CallBinary(BinaryFunc::CastTimestampTzToString)

would become:

    ScalarExpr::CallBinary(BinaryFunc::CastTimestampTzToString(<inlined tz>))

Determining whether a dataflow fragment can be shared now reduces to the
problem of determining whether two RelationExprs are equivalent.

The EvalEnv becomes a catalog/sql/coord-only concern, now called
PlanContext. If the plan results in creating a new item, then the
PlanContext needs to be persisted alongside the item, but otherwise the
PlanContext is entirely handled by the process of planning the SQL
query.