Support SHOW DATABASES { WHERE | LIKE }
Relevant to #1529.

This is an initial spike at support `SHOW ... WHERE` that only supports
SHOW DATABASES. Remaining work after this involves:
* supporting SHOW VIEWS ... which I think is fundamentally the same but
  involves reogranizing some logic.
* supporting other SHOW variants, some of which have multiple columns
  (I'm not yet sure how they choose what column to use for `LIKE`).

This uses a different approach than the existing implementations of LIKE
which I think is a little more generalizable: rather than running the
predicate on each row as the plan is generated, it just shoves all the
rows into a constant relational expression and puts a filter on top and
lets the execution engine (or more likely, the optimizer) sort out the
details.