testdrive: properly async-ify
Make testdrive use the async/await sugar, rather than manually
constructing futures and driving them around with a Tokio runtime.  We
don't actually care about or unlock the performance benefits of
async/await--this is test code, and not performance sensitive--but
because testdrive uses so many async components, using async/await is
far more ergonomic.

Hopefully this makes folks a lot less confused about how to use async
components in testdrive.

Note that the Avro OCF writer does not have an async library, and
presumably won't be the only testdrive dependency with no async support,
so a sync compatibility shim is provided for the actions that need it.