Don't track quote styles with identifiers
This commit simplifies (in some ways) the parser's handling of
identifiers.

Previously, to determine if an identifier needed to be quoted when being
printed out, the parser would track if it was quoted when it was read
in. This worked, but was a little clunky. After this change, we always
store the identifier in its normalized form (i.e., if it was entered
without quotes, it gets normalized, and if it was entered with quotes,
it does not) and determine if it needs quoting only when it gets printed
out.

This leads to a small change in behaviour in some cases, as the
capitalization of identifiers is now stripped away when it is not
relevant.