ci/builder: use cross-compiling toolchain to target old systems
We downgraded the CI builder image to Ubuntu Xenial (16.04) in order to
build a binary that was more widely compatible with old Linux
distributions. The rule is essentially you need to build on the oldest
distribution you wish to support.

Building on Xenial is annoying, though, because it contains relatively
ancient versions of various tools. Python v3.5 reaches EOL this year,
for example, and some Python tools already only support v3.6+. While
it's possible to install a more modern version of Python from a PPA,
this bloats the image over time, e.g. as other tools depend on the
distribution-provided Python.

Instead, we can build a cross-compiling toolchain that targets an
arbitrarily old version of the Linux kernel/glibc. This commit
introduces a cross-compiling toolchain that targets Linux 2.6/glibc
2.12, which means the built binaries will run even on a distribution as
old as RHEL 6. (Seriously, this was verified.)

The builder is then restored to Ubuntu 18.04, where Python 3.6 is
available.

A side effect of using this new toolchain was noticing that we were
failing to link libz statically. This commit also plumbs the correct
option to rust-rdkafka to remedy this.