testdrive: wait for shards after creating a stream, too
Ingesting immediately after creating a stream would occasionally fail
with a "stream does not exist" error. Likely the result of inconsistency
on AWS's end. Stick in a retry loop after creating a stream to wait for
the stream to become visible. This required a bit of refactoring to
share the very similar code with the update stream count action.