Introduce bin/mzconduct
The goal of mzconduct is twofold:

* It provides a central repository of workflow steps, that allow us to package complex
  workflows in a robust and introspectable way
* It introduces a uniform programatic api to execute any of our docker-compose tests
  * It makes it easy to verify that a process runs to completion, making it easier to run
    compose integration tests in CI

It should make it easier for all new integration tests to come up correctly with fewer
complex docker builds or hacks around their expected startup behavior.

It should make it easier to notice and faster to debug when steps *fail*, which is
something that is not necessarily obvious when every docker container is responsible for
its own dependency checks on startup.