coord: don't loop forever on erroring metadata requests
The timestamper could wedge forever in a Kafka metadata retrieval loop.
This would prevent any sources from getting timestamped, leading to the
system wedging. (This presented in CI as flakes in testdrive.)

In particular, if a topic is deleted from Kafka before the timestamper
is instructed to drop the source, the metadata requests on the topic
will fail forever. If auto topic creation is enabled, as it previously
was in CI, we would accidentally recreate the topic and break out of the
loop; but with that setting disabled, we get stuck forever.

Instead of assuming that every metadata request will succeed eventually,
bubble the error up so that it can be handled appropriately.