mzbuild: warn if there are files that have been deleted
mzimage and mzcompose will print a traceback if a file has been added to git and then
deleted, without doing a `git rm` or `git add <deleted file>`.

While it's possible that handling this in the error case so that errors are printed after
the traceback, if you _don't_ get an error I think the warning is even more important, in
case you thought that the thing being deleted was contributing to the fingerprint, but
actually is not.