implement as_of for sources (#2797)
This PR updates source creation to respect the as_of component of a DataflowDesc. This was previously used for index imports, but not for sources. This may not have been a problem, as it was not clear that you could create a DataflowDesc that had an as_of set and used bare sources, but it is becoming possible so we should fix it up.