sql: limit places where we parse wildcards
Wildcards are really special in SQL syntax, and can only appear
unqualifed in a select list, like

    SELECT * FROM ...

or inside of an aggregate function, like:

    count(*)

Our parser previously permitted wildcards anywhere in an expression
tree, which didn't make much sense. Adjust it so it only allows
wildcards in the two places mentione dabove.