dashboard-logs: Printing causes problems, so write to a file instead
By default in ubuntu docker containers writing to /dev/stdout is similar to printing, but
it doesn't share the same fileno as the actual stdout stream, which seems to mean that it
doesn't have problems locking stdout when sharing file descriptors with supervisord.

Additionally, because of the fact that FIFO files block their owning process if there is
not both a reader and a writer attached, we need to open the same file in both modes,
even though we're only reading from it.