ccsr: make config builder the only interface
It is unusual in Rust to have an object with a builder where the builder
has a build method that construts the object, and the object has a new
method that takes the builder. (If there is a way to create the object
with a default config, then a new/default method that does not take a
config is typical, but that's not the case here.)

This patch makes ccsr::ClientConfig::build the only means of creating a
ccsr::Client. It also adjusts the method to take ClientConfig by value,
as is conventional, to avoid clones in the common case. On the rare
occasion that the ClientConfig needs to be preserved, it can simply be
cloned by the caller.