mzbuild: don't pass --pull to docker build
If building an image where there is a FROM directive for an image that
is not available on Docker Hub, `docker build --pull` will fail the
entire build.

The point of `--pull` was to pick up changes to base images on Docker
Hub, but that's against the philosophy of mzbuild anyway, where images
are immutable. Instead, we can refer to Ubuntu's immutable tags
(bionic-20200403 instead of bionic, for example) and drop `--pull`. If
we need to update to a newer version of Ubuntu, we simply update the
version in the FROM line, and a normal `docker build`, without `--pull`,
will fetch the new version.