pgwire,materialized: rework connection dispatching
Introduce explicit HTTP and pgwire servers, which store configuration
parameters that are global to an instance of materialized. (Previously
these parameters were passed through as function arguments for every
connection, which was getting unwieldy.) This will be used in a
forthcoming TLS patch, which will teach the HTTP and pgwire servers to
accept TLS connections.