materialized,testdrive: include kinit and kdestroy in Docker images
materialized now statically links libkrb5, but an important part of
Kerberos is the kinit binary, which obtains the initial
ticket-granting ticket and is invoked automatically by librdkafka.  This
binary is normally supplied by the krb5-user package, or the equivalent
on non-Debian systems, but since we're statically linking libkrb5 we
need to provide our own statically-linked copy of kinit.

It is in theory possible to use the distribution's kinit with our copy
of libkrb5, but this can lead to all sorts of subtle problems if the
distribution's kinit is from a significantly different version of
libkrb5 than Materialize was linked against.

For good measure, this commit includes kdestroy, since it is the
complement of kinit. Additional binaries can be added if they prove to
be necessary, but for now we'll limit it to these two to reduce the size
of the resulting Docker image. (Each binary is ~2MB in size.)

Fix #2869.