python: force-disable pyyaml's c extension
By default PyYAML will check to see if it can compile a C extension that
links against libyaml. This is just another thing that can go wrong when
setting up the venv, and we don't need the additional speed of libyaml
over the pure Python YAML parser, so just force disable libyaml
extension.