Thread common state through all transforms
We need access to an IdGen to write certain transforms. Take this
opportunity to bundle up all the state that a transform has access to in
a struct. In the future there should be some kind of "global state" that
is common to every stage of planning, that provides everything necessary
to construct expressions.