materialized: refactor server shutdown
@quodlibetor remarked that the use of std::sync::Weak in server shutdown
was a bit non-obvious, so this commit is an attempt to refactor that
confusing bit of code away.

Instead of the pgwire and http servers checking whether they can upgrade
their weak cmdq_tx on every new connection, they get a normal cmq_tx.
When it is time to begin shutting down the server (a process now called
"draining"), the pgwire and HTTP servers are dropped in their entirety,
including the contained cmdq_txs.

We introduce a new dependency on the stream_cancel crate, which is
designed for exactly this purpose, to communicate the drain message
across task boundaries.

This also fixes a resource leak present in the old code, where the
listener task would never be cleaned up until the process exited. This
wasn't a problem in production, because there is no way to terminate the
server without exiting the process, but it would cause TCP listeners and
Tokio threads to pile up when running unit tests.