ci: add --pretty flag to mypy
It causes mypy to point to the source of the error (similar to rust error messages)
instead of just listing the file/line.

Note that we _don't_ want this in mypy.ini because that breaks the error introspection of
tools like emacs/vim that shell out and expect `path/to/file:line:col` syntax.