Better respect compaction frontier (#3005)
As described in #2803 we largely ignore the since compaction frontier of arrangements, most especially in the creation of dataflows for new arrangements and sinks. This PR forces that information in to dataflow construction and ensures that dataflows are constructed with such a minimum frontier and that newly created arrangements use this frontier as their starting value for their compaction frontier.