Push more filters down via inlining
This commit "absorbs" sufficiently simple Map expressions referenced in
Filters in order to push them down. In order to avoid exponential
blowup, this is only done in limited cases: only unary functions and
binary functions with a leaf input, and the expressions must only
reference input columns.

I think the last restriction could be lifted, but the appropriate way to
do so would probably be to inline those expressions in the Map itself
and subsequently let this transform trigger.

I also added a directory in the slt directory specifically for testing
transforms like this.