Add a simple Python compilation check to the lints
With Python syntax errors mypy is capable of giving very bad error messages, I suppose
specifically with things in f-strings.

With this change, error messages may get duplicated, but they also may provide actual
information. Specifically, this makes the error from the mzconduct intro go from:

    misc/python/bin/mzconduct.py:1: error: invalid syntax
        # Copyright Materialize, Inc. All rights reserved.
               ^
    Found 1 error in 1 file (checked 22 source files)

To:

    misc/python/bin/mzconduct.py:1: error: invalid syntax
        # Copyright Materialize, Inc. All rights reserved.
               ^
    Found 1 error in 1 file (checked 22 source files)
    *** Error compiling 'misc/python/bin/mzconduct.py'...
      File "<fstring>", line 1
        (error=)
              ^
    SyntaxError: invalid syntax

It's not perfect, but it's not completely incorrect anymore, and it at least includes
something you can search for.