wip: introduce a testing framework for transforms
In general so far, I've found it pretty hard to write tests for specific
transforms in the transform crate, and I think the volume of tests in
there indicates that I'm not the only one. This PR is my attempt to fix
that. I'm curious to hear if people think this would or would not be
useful, or if they have other ideas for how we could make these tests
easier to write.

My proposed solution is this: a small new frontend to the IR whose
semantics map more-or-less identically onto it.

I'd happy to bikeshed the syntax a bit (though perhaps not the
implementation yet), but I think something like this should make it much
easier to write targeted tests for transforms.

This framework supports general rewriting via an env-var:

```
REWRITE=1 cargo test -p transform
```

will regenerate all the tests files, much like how the sqllogictests
work. However, unlike the sqllogictests, this has some benefits:
1. not required to go through the SQL layer to test,
2. allows us to run any subset of the transforms on a given IR that we
   like,
3. rewrite support means it's feasible to have tests depend on the
   entire set of transforms.