python: stub make materialize installable in a standard way
This makes the materialize project a bit more like a standard Python project, in the
interest of making it play nicer with virtualenv-management tools.

It is possible to now install most of python/materialize (not the scripts) using:

    pip install misc/python

or:

    pip install "misc/python[dev,ci]"

This also adds the `py.typed` file, which makes mypy happy if you don't run it with the
exact runtime configuration of `bin/pycheck`.

I care about being able to use the virtualenvwrapper and tox family of projects,
partially because all my tooling is built around them, but also because they make
tearing down/bringing up virtualenvs easy and this brings us a bit closer to being able
to test that more casually.