Fix ANY/ALL regression
This commit fixes a bug introduced by #3045 in the presence of NULLs: if
there is a NULL present in an ANY that would otherwise be false, the
result is NULL, and likewise with ALL and true.

Since we want to retain the planning benefits, we do a sort of hacky fix
where we require the right column to not be nullable to make use of the
EXISTS transformation.