Fix another problem with ANY/ALL => Exists
The original change also didn't handle the case where the LHS was
nullable. This is sort of a hacky fix at the moment, we just construct
an `if`, but we should be able to eliminate it in many cases.

This isn't an ideal fix, since it duplicates the possibly expensive LHS,
but it doesn't seem to practically hurt in any tests we have today.