py: Add pytest execution to `pycheck`
This takes it a bit further from `cargo check`, perhaps pytest should be run in a
separate thing called `pyvalidate`? As it stands right now the only test I'm adding (in
the next commit) runs in less than a 1/2 of a second, including python startup, so I'm
not sure if it's worth the extra mental complexity of having different tools for folks.

Happy to add it, though.