doc/user: fix and lint for broken links
This patch adds a CI jobs that uses htmltest [0] to lint for internal
broken links and images. This exposed a number of links that were
broken due to various docs refactors, all of which have been fixed in
this patch.

This commit additionally makes use of Hugo's "markdown render hooks" to
automatically make links and images of the form `/path` relative to the
site's base URL. For example, `/images/foo` will be rewritten to
`/docs/images/foo` when deployed to production.

[0]: https://github.com/wjdp/htmltest