sql-parser: support general typed string literals
Support PostgreSQL's typed string literals, as in:

    SELECT bool 'true'

Previously we only supported the typed string literals for the date,
timestamp, and timestamp with time zone types as required by the SQL
standard, but PostgreSQL generalizes this syntax for all types.

As a side effect, this fixes the bug where

    SELECT date

parsed as a syntax error rather than as selecting a column named "date".
The fix also generalizes, e.g., `SELECT bool` is also valid.

Fix #2577.