repr,expr,sql: remove ScalarType::Unknown
ScalarType::Unknown is a Postgresism that has not served us well. It
is best understood as the type of NULL in the following SQL statement

    SELECT NULL

although in recent versions Postgres will assign that NULL a type of
`text` before planning is complete. That is in fact the key insight: the
"unknown" type represents a transient state of the planner. Either the
unknown type will become known by the time planning is complete, or it
will not, and planning the query will fail.

We can isolate the dataflow execution layer from the unknown type by
limiting its existence to the few moments in the planner where it is
needed. To that end, this patch removes the ScalarType::Unknown variant,
and introduces a CoercibleScalarExpr to take its place.

The idea is that the planner has special handling for expressions which
could possibly coerce to another type (i.e., a string literal, list
literal, NULL literal, or query parameter). The expressions will be
coerced at the moment their true types are required; for example calling
a function will coerce any unknown parameters, as will reaching the root
of the expression tree.

All code outside of the `sql` crate gets vastly simpler. The SQL planner
gets simpler in many places too, and a bit different in others--not more
complicated, necessarily, just different.