Add an mzconduct buildkite plugin
The next few commits will start using it, some changes are necessary to tests and demos
to use it safely.

This almost completely reuses the shutdown step from mzcompose, since we really want to
just forcefully kill things, but it does add features to mzconduct to make it integrate
reasonably.

This also leaves the mzcompose plugin, because we do seem to have several workflows that
don't do just start up and run correctly with no ordering guarantees.