sql: remove DatabaseSpecifier::Temporary
With Catalog::get_schemas out of the picture, we can complete #3104 and
use (DatabaseSchema::Ambient, "mz_temp") to represent the temporary
schema. This is a bit more semantically correct, as "mz_temp" is
logically part of the same ambient database as mz_catalog and
pg_catalog. It also simplifies the storage of temporary schemas in the
catalog, and exposed a bug where `SHOW EXTENDED SCHEMAS` was not showing
the temporary schema.

Fix #3104.