sql-parser: clean up API
We inherited a lot of cruft from upstream on the SQL parser API, which
tries to be much more general than we need to be. Hide away almost all
the public API surface, leaving only the AST and two free functions,
parse_statements and parse_expr, exposed.

Since SQL parsing is stateless to the outside observer, hiding away the
Parser object that tracks state while parsing is a nice simplification.