dataflow: don't block on Kafka metadata refreshes
Synchronously fetching metadata in the Kafka source causes problems. At
best it blocks the worker thread from performing computation for a few
milliseconds. At worst it blocks the worker thread for the full 1s
allotment and then times out, resulting in a second of lost CPU time and
a wedged Kafka source. Our hope was that production Kafka clusters would
be reliable enough that the synchronous fetch would be a rare problem,
but empirical evidence has shown otherwise.

It turns out there is no need to synchronously fetch metadata in the
Kafka source. librdkafka is happy, almost too happy, to allow us to
subscribe to a partition for which it does not have metadata. The only
stipulation is that we will not receive any data for the partition until
the partition's metadata is retrieved.

There are two main ways that the metadata refresh occurs. The first is
an explicit call to `fetch_metadata`. The second is the
regularly-scheduled background metadata fetch, controlled by
topic.metadata.refresh.interval.ms, which defaults to 5m. This patch
uses a combination of both strategies for optimal performance
characteristics.

First, this patch removes the synchronous metadata fetch entirely. Upon
arrival of a new partition ID, the Timely worker responsible for that
partition blindly assigns the consumer to that partition. If the
partition ID is invalid, the source will wedge, but a) the partition ID
can't be wrong with real-time timestamping, and b) if you use BYO
consistency, you've put us in an impossible situation, so wedging is
possibly the most correct outcome.

Second, this patch turns sets the background metadata fetch interval to
30s, down from its default of 5m. I suspect this is entirely sufficient
for production deployments, where adding a partition is a rare operation
on a human timescale, but it means our tests would take 30s to detect
new partitions.

So, finally, this patch triggers an asynchronous metadata fetch whenever
it assigns a partition. This means tests detect new partitions almost
instantly. We don't actually need to wait for the metadata fetch to
complete, and it's not a problem if it fails, since the background
metadata fetches will retry indefinitely, and we just need one to
succeed eventually.

Note that there is a bug in librdkafka that generates spurious error
messages with this approach, but I've opened a patch to fix that. [0]

[0]: https://github.com/edenhill/librdkafka/pull/2916