sql-parser, sql: special case coalesce
COALESCE is a really special function because it coalesces the types of
its arguments in the same way that UNION, VALUES, and other bits of SQL
syntax do. PostgreSQL special cases COALESCE in the parser, so this
patch does the same.

This essentially moves a bit of complexity out of the general function
selector and into the main SQL planner, where it fits in a bit more
naturally next to all the other weird corners of SQL.