py: Fix a useless format trick
The say function allowed you to specify format args as inline parameters because it's a
pattern that I'm used to from before f-strings. However, with f-strings we literally
never use it, and it also can cause problems if you want to print a string that has curly
braces in it, for example because you're printing a dict.