coord: run realtime kafka metadata fetches in background
Some production Kafka clusters are known to take upwards of 1s to
respond to metadata queries. This means the 1s timeouts our timestamper
uses by default are too low, and can result in the timestamper never
making any progress.

For just realtime Kafka sources, move the timestamp metadata queries to
a separate thread, one per source. In time we can be much smarter about
reusing threads for multiple topics on the same broker, and apply a
similar approach to other sources (e.g. Kinesis, BYO Kafka) that run on
the main thread, but for now this solves the immediate problem.