move filter_lets to predicate_pushdown (#3273)
The filter_lets transform used to be where we push predicates down across Let and Get statements, hoisting common predicates from Gets in a Let body. In the interest of inter-view optimization, it makes more sense to have one method that can also collect candidate predicates for each Get statement, to push to other views or to sources.