sql-parser: reorganize AST for maintainability
Split the AST into more modules, so it's easier to navigate. Isolate the
types that are formally part of the AST from helper structs, so that
walkabout has an easier time finding the structs it cares about.

This commit is purely code movement with the exception of removing the
FromStr implementations from a few types (ExtractField and
WindowFrameUnits), as AST nodes should not have FromStr implementations.
It doesn't make sense to try to parse an AST node from a string without
invoking the full machinery of the SQL parser.