sql: complete pseudotype support for function selection
Before this change, functions that needed to resolve competing implementations
when called with parameters would encounter behavior that wasn't aware of
pseudotypes, which could lead to failing to resolve a function that we should be
able to select. This change teaches that section of the function selection
algorithm about pseudotypes.