sql: don't derive Clone for FuncImpl
This will allow us to use a boxed closure as the operation type. The
interesting bit here is the integration of `saturate_decimals` into
`coerce_arg_to_type`: rather than rewriting the function parameters to
have the correct decimal precision and scale, we just avoid planning a
cast if the function desires a decimal and we already have a decimal.