bin/xcompile: fix cross-compilation
Remove the environment variable overrides in bin/xcompile that apply
globally, like `CC` and `AR`. Compiling Materialize involves compiling
some programs for the host as well as the target, and `CC` applies to
both compilations; setting it indiscriminately to
x86_64-unknown-linux-gnu means that these host programs are miscompiled.

Also, remove the download and installation of libz. We statically link
libz now so there is no need to acquire a dynamic version of a Linux
libz.