Cause certain Kafka metrics to be deleted when their resources are
This uses MaterializeInc/rust-prometheus#2 to guarantee that at least some metrics stop
being reported when `DROP SOURCE` is performed. This is enough to resolve the false
metrics appearing in #3352, but it does not delete all Kafka metrics when their sources
are deleted, in particular in the timestamper, where we don't have structs that live
along with their partitions, yet.