Do not hoist expressions from `if` bodies (#3395)
Scalar CSE was hoisting common expressions in then and els expressions, without certainty that they would be evaluated. This led to queries like
```
SELECT
    CASE WHEN b = 0 THEN 0 ELSE 1/b END,
    CASE WHEN b != 0 THEN 1/b ELSE 0 END
FROM x
```
unconditionally evaluating 1/b, because it would be hoisted prematurely. We don't do that now. We could be more advanced and determine the expressions in the intersection of the then and els expressions, but that is future work for now.