interchange: fix Avro record field name invention
The Avro encoder was going to a good bit of trouble to invent field
names for unnamed columns, but it was accidentally building the Avro
schema before it had invented the column names. Reorder the steps to fix
the bug, and tweak the types inovled to avoid such bugs in the future.

Also, tweak the column numbering to start from 1, to be consistent with
other places where we invent column names.