expr: reorganize scalar functions
The expr::scalar::func module had gotten out of hand, as it contained
all scalar functions in one file. This patch splits that module into
thematic submodules, where functions are grouped by the primary data
type they operate on.

This patch additionally restructures how we track properties of
functions (e.g.  whether a function propagates nulls or preserves
uniqueness). The old approach worked like this:

    impl BinaryFunc {
        fn is_some_property() -> bool {
            match self {
                BinaryFunc::Foo | BinaryFunc::Bar => true,
                _ => false,
            }
        }
    }

This was quite error prone. For one, it was quite easy to forget to
update `is_some_property` when adding a new function. (Exhaustive
matches could help here, and indeed were used in a few places, but
making everything an exhaustive match would have made an already large
file unnavigable.) Perhaps more problematically, the specification of
the properties was separated from the implementation of the function by
hundreds of lines, and so it was easy to specify the properties
incorrectly, resulting in subtle bugs like #3403.

The new approach involves filling out a `FuncProps` struct near the
implementation of each function, like so:

    FuncProps {
        can_error: false,
        preserves_uniqueness: false,
        nulls: Nulls::Sometimes {
            propagates_nulls: true,
            introduces_nulls: false,
        },
        output_type: OutputType::Fixed(ScalarType::String),
    }

We use Rust's requirement that all fields of a struct be initialized to
enforce that all properties are explicitly spelled out. (If this becomes
onerous, we can always introduce some helper/builder methods.) This
makes it impossible to forget to specify a property when adding a
function, and the fact that the FuncProps struct is near the
implementation makes it harder to get the properties wrong.

For extra safety, this patch also includes a dynamic property checker
that runs in debug mode and fails if a function evaluation produces a
value that is inconsistent with its properties. This isn't a perfect
solution, since the checker works at runtime and can only catch
violations that actually occur, but it's much better than the status
quo.

Also new in this patch is the `introduces_nulls` property, which
distinguishes between functions that can *create* nulls from non-null
inputs and those that merely propagate nulls. This will be used in a
forthcoming optimization of the IsNull function.

Besides the new property, which is inert in this commit, there are not
intended to be any semantic changes in this commit.

Fix #1379.