Change how the coordinator assigns timestamps
Based on some discussions with Frank. Still kind of a WIP, since there's
a handful of things I'm unsure about.

First of all, it seems sketchy that there's no guarantee that a
timestamp we assign for a read will get closed in the future, but
perhaps in practice that's not a problem.

Second of all, this seems to slow down the running of the logic tests,
on my laptop they go from 12s to 17s, and I'm not really sure of the
culprit, but posting this here in case there's anything glaringly
obviously wrong.