expr: improve reduction of IS NULL
An expression like `(a + b) IS NULL` can be reduced to `a IS NULL OR b
IS NULL` because we know the `+` operator can never introduce a NULL.
This can unlock further reduction, e.g., because `b` is known to be
non-NULL.

This fixes the regression described in #3488.

Tracking this "introduces null" property requires yet another method on
BinaryFunc which is hard to maintain. The fix is a massive refactoring;
see #3449 for a sense fo the scope.