Introduce a deterministic order to enumerate dataflow imports. (#3453)
We were using hashmaps to determine the order in which we imported sources and indexes into dataflows. This can be non-deterministic, so we shouldn't do that. This manifested in #3448, and it is possible that with single-process the hashmaps were effectively cloned, whereas when serialization gets involved they naturally end up in different maps with different random seeds.