sql-parser, sql: hoist AST transforms into transform_ast module
Various bits of SQL syntax are just sugar. For example,

    <expr> BETWEEN <low> AND <high>

is just sugar for:

    <expr> >= low AND <expr> <= high

`IN`, `ANY`, and `ALL` are similar. Hoist our handling of this syntax
sugar into a new "Desugarer" transformer in the transform_ast module,
which performs an AST -> AST transformation to rewrite the syntax sugar
in terms of the more fundamental syntax.

This patch is more than just cosmetic; it paves the way for record types
(#3376) by ensuring that comparison operations are directly visible as
`<lhs> <op> <rhs>` and not hidden behind several layers of sugar, so
that comparisons between ROW constructors can be properly discovered and
planned.