ccsr,sql: support SASL PLAIN authentication for Kafka
Support the SASL PLAIN authentication scheme for Kafka and the HTTP
basic authentication schema for the schema registry. This makes it
possible to use Materialize against a Confluent Cloud cluster.

Note that determining which subset of Kafka configuration options are
allowable for a given security protocol is approximately impossible, so
instead we just dump all the options that are allowable in *some*
configuration into one big allow list and let librdkafka sort out the
rest.

Fix #3418.