Avro decoding rewrite (#3568)
This diff accomplishes two main goals:

(1) It rewrites avro::decode and interchange::avro to decode directly into rows, rather than into intermediate avro values. This significantly speeds up decoding (3-4x).

The API is loosely inspired by Serde: a "Deserializer" walks the schema, and calls functions provided by a "Decoder".

Unfortunately, when fields do not appear in the writer in the same order as they are expected by the reader, we need to fall back to the old API, because we need to reorder the values before writing them to the row.

(2) It allows arbitrarily nested records and lists to be ingested into Materialize.