sql: allow input columns in ORDER BY clause
In standard SQL, the ORDER BY clause occurs after the columns specified
in the SELECT list have been projected. In a query like

    CREATE TABLE (a int, b int)
    (SELECT a FROM t) UNION (SELECT a FROM t) ORDER BY a

it is valid to refer to `a`, because it is explicitly selected, but it would
not be valid to refer to unselected column `b`.

But PostgreSQL extends the standard to permit queries like

    SELECT a FROM t ORDER BY b

where expressions in the ORDER BY clause can refer to *both* input columns
and output columns.

This commit adds support for this extension. It also fixes a
longstanding bug in RowSetFinishing::is_trivial that was exposed by this
change.

Fix #696.
Fix #2260.