coord: change more logging views to consistent catalog views
This commit moves the following logging views over to the new catalog
view framework:

  * mz_avro_ocf_sinks
  * mz_catalog_names
  * mz_view_foreign_keys
  * mz_view_keys

These views are often queried immediately after a catalog-altering
operation, so it makes for a much better user experience if they are
guaranteed to be up-to-date with respect to catalog changes. For
example, we are able to remove several retry loops from our sink test
code, instead relying on the newly-guaranteed consistency of these
views.

Only one usage of SequencedCommand::AppendLog remains, for frontier
information. Porting this over to a catalog view is tricky, as it
results in runaway CPU usage: if it's a catalog view, updating a catalog
view results in new frontier information, which results in more updates
to the catalog view...

Fix #2576.