sql-parser: permit table functions to be preceded by LATERAL
PostgreSQL permits table functions to be preceded by the optional noise
word LATERAL, as in:

    SELECT * FROM foo, LATERAL bar()

This does not change the semantics of the query, as table functions are
always implicitly planned as if LATERAL was specified. This commit
teaches the parser to accept this syntax, but not to retain the
distinction.