sql: properly special case aliases for single-col table funcs
Add a strange special case for table functions that ouput one column for
PostgreSQL compatibility.  If a table alias is provided but not a column
alias, the column implicitly takes on the same alias.

Concretely, this means `SELECT x FROM generate_series(1, 5) AS x`
returns a single column of type int, even though

    CREATE TABLE t (a int)
    SELECT x FROM t AS x

would return a single column of type record(int).