Render joins even if arrangements are missing (#3726)
Rendering of joins would previously panic if the join asked for an arrangement that was absent. Because some optimizations fight, this wasn't always the case (constant expressions would be hoisted out of the ArrangeBy operator). We certainly do want to ensure that there is a clear spec for the RelationExpr AST, but also we shouldn't panic when we don't need to.