comm: disable maximum frame length
comm channels are trusted, so we don't actually want a maximum frame
length at all. (Whatever we pick is destined to be too small; someone
somewhere has a machine where shipping around N gigabytes is reasonable,
for all values of N.) Just disable the frame length check entirely. If
we can allocate it, it's ok.

This isn't to say that we shouldn't be more helpful in preventing folks
from OOMing their Materialize instance, but enforcing limits at the comm
layer is pretty unhelpful, because the limits are not tied to anything
the user understands. Better to enforce limits on e.g. the maximum
number of rows, or the maximum pgwire result set size, rather than
"result sets must not exceed 1GB when encoded using Materialize's
internal bincode representation." Fixing that in a principled way is
punted to another commit.