Decorrelate conditional subqueries carefully (#3736)
This PR makes decorrelation more robust for subqueries in conditional logic (the then and else cases for ScalarExpr::If). The idea is that if no subqueries exist in either case, we use the standard approach (decorrelate each expression, return the If), but if subqueries exist in either then or else, we independently decorrelate each case having first filtered by cond appropriately. For each filtered and decorrelated relation, we apply the logic as a map and change the expression to look at the last column.