dataflow: split render::build_dataflow into chunks
There is still far too much complexity in rendering, but this makes it
possible to reason about each step of the dataflow build more
independently.

There are no code changes in this PR, except that logging of dataflow
dependencies is hoisted out of the render module, to simplify the amount
of state that must be plumbed.