dataflow-types: don't key sources by SourceInstanceId
Unclear what was going on here. DataflowDesc::source_imports wants to be
a map from (imported source ID) -> (imported source description); it has
no business or interest in importing a particular *instance* of a
source. And indeed, no one was interested in the anything but the "sid"
field of the SourceInstanceId, which makes this change a strict decrease
in incidental complexity.