expr,sql: give every aggregate function an identity
In order to transform a SQL expression of the form

    <agg>(<expr>) FILTER (WHERE <cond>)

into

    <agg>(CASE WHEN <cond> THEN <expr> ELSE <identity>)

every aggregate function must have an identity. The transformation was
previously assuming that null was the identity for every aggregate
function, which was wrong. For example, the identity for ANY is false,
not null.

There are two big changes in this commit:

  * AggregateFunc::CountAll is removed. There was *no* identity for
    this function, because it did not look at its input. Its usage is
    replaced by Count(true).

    This has the potential to make count(*) a bit slower, but I can't
    imagine the extra null check being that expensive in practice.

  * AggregateFunc::JsonbAgg learns to treat null as the identity.
    These are not the semantics that SQL wants, but the SQL layer can
    easily recover those semantics with a coalesce shim. (Shimming
    an identity onto the old semantics, by contrast, is not possible.)

Fix #3872.