sql: don't elide identity joins when LATERAL
A query of the form

    SELECT ... FROM t1, LATERAL (SELECT * FROM table_func(t1.col))

would previously cause a panic in decorrelation. This query plans as,
roughly:

    %1 = Get t1
    %2 = Constant ()
    %3 = CallTable table_func(#^^0)
    %4 = InnerLateralJoin %2 %3
    %5 = InnerLateralJoin %1 %4

The planner was previously being too clever and, noticing that %2 was
the join identity, attempting to simplify to:

    %1 = Get t1
    %3 = CallTable table_func(#^^0)
    %5 = InnerLateralJoin %1 %3

And, indeed, this is almost valid... except that the column reference in
%3 is still doubly outer, when it needs to be only singly outer.

We *could* fix up the column references, but this is a lot of work for a
normalization rule that exists only to try to tame the output of
`EXPLAIN RAW PLAN`. (The optimized plan is the same regardless of
whether we apply this rule during SQL planner). So, instead, just don't
perform such elision on lateral joins.

Fix #3877.