doc/user: add simple version shortcodes
Add some simple version shortcodes that make it easy to indicate when a
feature is still experimental, or when it is on master but not yet in a
release. This is a quick fix for not having a more thorough story around
versioning the docs.

To demonstrate the approach, I retroactively added version shortcodes
for features added or changed in v0.4.0, but I don't plan to do this for
earlier versions.

Fix #3868.