coord: properly handle duplicate source instances in persistence
Previously, it was possible for multiple source instances to send copies of
their data to the persister thread and the persister thread would write out
multiple copies of the data to disk. This commit changes that by deduplicating
messages by offset before writing to persistent storage, and changes the behavior
to keep the copy with the minimum timestamp over all duplicates.

Alternatively, we could have changed the structure of the persister thread to
key sources on SourceInstanceId instead of GlobalId, but that would mean that

1. We would write down multiple copies of the source data
2. It made the restart path a lot more complicated, as now the coord layer needs
to know a source's SourceInstanceId before it can determine what data is
available to be read.