coord: don't track view uses separately from catalog
We were duplicating the dependency information in both the coordinator
metadata and the catalog. This added complexity and the possibility of
the metadata getting out of sync. Rework so that the coordinator just
uses the view dependency information from the catalog, rather than
duplicating it.

Note that the comment that said "Only views, not sources, on which the
views depends" appears to have been entirely wrong.