coord: derive queryability information from catalog
Whether a view is "queryable" or not derives from whether it is possible
to find indexes for all of the view's inputs. We were previously
denormalizing this information in the coordinator and keeping it up to
date by manually propagating querability when new indexes were created.

There were two big downsides to this approach:

  * The logic was a bit brittle and error prone. Modifications to the
    coordinator could easily result in missetting the queryability state,
    which caused hard to track down bugs.

  * The metadata only existed in the coordinator, and so `SHOW VIEWS`
    and `SHOW SOURCES` had to be special-cased in order to be computed
    in the coordinator. Besides being a bit annoying, this was about to
    impede #913 (mirroring `SHOW` data as views), and prevented using
    `SHOW {SOURCES,VIEWS} ... WHERE`.

The new approach simply computes the queryability property on demand
using information that is available in the catalog, resolving both of
the above downsides.

One potential downside of the new approach is that `SHOW VIEWS` has to
do more work to compute queryability, but the performance of `SHOW
VIEWS` is not particularly important. The hot path (`SELECT`) is
probably unaffected by this patch, since computing queryability is done
concurrently with finding the nearest indexes for the view.