expr: allow unioning an arbitrary number of inputs
RelationExpr::Union was previously limited to unioning two inputs at a
time. This was problematic for VALUES clauses in particular, which are
planned by unioning together many single row collections. For example

     VALUES (1), (2), (3)

is planned as:

    %0 =
    | Constant ()
    | Map 1

    %1 =
    | Constant ()
    | Map 2

    %2 =
    | Union %0 %1

    %3 =
    | Constant ()
    | Map 3

    %4 =
    | Union %2 %3

This resulted in very deep trees of union operators with large VALUES
clauses. Attempting to plan these deep trees would lead to stack
overflows, as in #3995.

Since large VALUES clauses are common, teach RelationExpr::Union to
support an arbitrary number of inputs. Now the example above plans as:

    %0 =
    | Constant ()
    | Map 1

    %1 =
    | Constant ()
    | Map 2

    %2 =
    | Constant ()
    | Map 3

    %3 =
    | Union %0 %1 %2

This scales much better, and avoids overflowing the stack. (It still
plans + optimizes slowly, but we have to start somewhere.)

Multi-unions are also likely to be slightly faster in dataflow
execution, since they plan as a single timely concat operator, rather
than a nested tree of concat operators.

Fix #3995.