prof: merge back into one crate
The split between prof-common and prof-jemalloc is awkward, and exists
only to work around Cargo's lack of target-specific features. Work
around that in a different way, so that prof can be a single crate whose
jemalloc support is conditional on a "jemalloc" feature.

The new workaround involves depending on a special
"materialized-not-macos" crate when building for non-macOS targets. This
crate enables the jemalloc feature of the prof crate, and due to the way
Cargo unions features together across crates, this enables the jemalloc
feature of the prof crate for *everyone*. This is essentially a hack for
getting target-specific features on stable Rust.

There are two big advantages of doing it this way:

  * `prof` is closer to being ready for public consumption.
  * The approach generalizes to future target-specific features that may
    need to be enabled, especially if those features are in dependencies
    which we do not control. (We could only split up the prof crate into
    its macOS and non-macOS halves because we controlled it.)