misc/tb: absorb "tail binlog" (tb) program
tb is becoming increasingly important for our customers. It previously
lived in a separate repository

    https://github.com/MaterializeInc/tb

where it was not subject to CI or our coding standards. This worked well
when tb was a proof of concept, but it is now more important than that.

This patch absorbs tb into the main Materialize repository. Immediate
benefits include:

  * A new integration test that uses mzconduct.

  * Integration with mzbuild so that the materialize/tb Docker image is
    automatically kept up to date.

  * Adherence to the MaterializeInc/materialize code review standards.

The big downside is that it is a bit harder to contribute to tb
directly, but I think the tradeoff is well worth it.