tb: refactor + support resuming from checkpoint after restart
The headline feature here is support for resuming from the committed
checkpoint when tb restarts without rescanning the whole source
database. This doesn't guarantee that there won't be duplicates, but
Materialize should in theory be able to handle that deduplication.

This work includes a major refactor of tb as well. I apologize for
refactoring and changing behavior in one commit, but the whole of tb is
small enough that I hope it's not too hard to review. The new code
avoids mixing a bunch of static and non-static state together in one
large class, so it should be a good bit easier to follow.

Fix MaterializeInc/tb#1.