testdrive: fix validate_sink and consolidation.td
Previously, validate_sink would not catch an error when comparing a single expected
record against a empty vec of actually received records, because the function would
zip two iterators together and one would be advanced, while the other was not. This
commit fixes that behavior by calling next() on both iterators in parallel and
complaining loudly if only one of them returns None.

test/testdrive/consolidation.td was broken in a few ways. First, the CREATE SOURCE
command and the kafka-ingest commands used incorrect Kafka topic names for the
consistency topic and the data topic respectively. In both cases, we forgot to
include the "testdrive-" prefix and "-${testdrive-seed}" suffix on topic names.

Furthermore, the timestamping code expects to get Debezium formatted consistency
information for ENVELOPE DEBEZIUM sources, and silently fails to parse the
"Materialize-native" format in that case.