coord: augment source connectors with previously persisted files on-demand
Previously, we only checked to see if we could reuse already persisted data
after restarts, and we would then cache that information in the catalog. This
was a simple behavior but ultimately not very useful. This commit teaches the
coordinator to augment source connectors whenever they are being imported
into a dataflow, which opens up more opportunities for the persisted data to
be useful, and makes persistence easier to test.