pgtest: add postgres wire protocol tester
Add a postgres wire protocol tester. This is useful because it allows
us to test things against a real postgres server and compare the results
to mz without needing to depend on drivers and their specific behavior.

To demonstrate this, a test file is included that demonstrates a current
bug in our pgwire implementation, hence why it isn't hooked up to
builds yet.