Add a table function to view persisted records
A little loose at the moment since I think it's a bit of a pain to
thread through the configured persistence directory, but should enable
some amount of testing. I tried to factor out the behaviour here as much
as I could but it was a bit of a pain. I'm a bit unsure about the
package boundaries in general here.