sql: fix literal lifting through unions
Previously the logic for determining the literals to lift through a
union was wrong, because it special-cased the `base` input, and could
result in changing the shape of some of the inputs in an invalid way.

This change changes this behaviour to extract the longest common suffix
between all arms, which makes sure that the shape of the inputs is
always correct.