materialized: don't create Tokio runtime in materialized::serve
Make it the caller's responsibility to construct a Tokio runtime before
calling materialized::serve. This affords the caller more flexibility,
and also makes the code inside of materialized::serve more idiomatic, as
it can use `.await`.

This is part of a commit series to improve our async/sync glue code.