comm: discover async executor from environment
Rather than requiring that the Tokio runtime handle be passed in
explicitly, just derive it from the environment. This is more in line
with how the rest of the Tokio ecosystem works, and makes code a lot
simpler.

This is part of a commit series to improve our async/sync glue code.