coord: make coordinator use async/await
It is more understandable for the coordinator to be composed of async
functions that periodically call the `await` operator, rather than
spinning up a separate futures executor via the `block_on` function.

Now, the only bit of incorrectly-synchronous code lives in the
timestamper module.

This is the last part of a commit series to improve our async/sync glue
code.