comm: attempt to deflake test_mpsc_connection_reuse
This test is sometimes hanging forever in CI. I'm unable to reproduce
this failure locally, but a common cause for failures of this sort is
using Tokio's basic runtime instead of its multi-threaded runtime.
Successfully using the basic runtime requires being very careful to not
require any actual parallelism, since everything runs on only one
thread, and likely this test is not that careful.