catalog: include keys when creating Builtin::Table Indexes (#4330)
Previously, the code creating and inserting indexes for our builtin system tables was
erroneously creating indexes with no keys. This meant that the indexes for the system
tables were not being inserted into the mz_indexes system table.

This change updates the definitions of the system table indexes to use their default
keys, instead.