materialized: add a SQL-over-HTTP interface
This will be used for an upcoming memory visualizer HTML page. It should
also be useable for a possible "type SQL into a box and see the results"
page.

Implement this by sending a SQL request to the coordinator from the
HTTP server. We avoid users, logins, and sessions by using a system
user which allows this to be stateless. From a security perspective,
this is identical to the (lack of) controls we already have.