sql: adjust type computation interfaces
The SQL planner deals with both "coercible" and "coerced" scalar
expressions. The main difference is that the type of a coerced
expression is known, while the type of a coercible expression
is possibly unknown.

The ExprContext provides helper methods (column_type and scalar_type) to
determine the column type and scalar type, respectively, of expressions.
Until this patch, column_type could be called on either type of
expression, while scalar_type could only be called on coerced
expressions. Fix this wart by allowing both column_type and scalar_type
to take either type of expression and return either a ScalarType or
Option<ScalarType> as appropriate.