www/public: add static memory vizualization page
Memory graphs now present a share link that uses our static
site. Information is passed via the URL hash as a JSON object. This will
allow for it to be extended in the future. The hash is used instead of
search ('?search') because the hash is not passed to the server by the
browser. This reduces privacy issues and also avoids any URL limits
by netlify (browsers may impose their own limits, but they are likely
much higher).