materialized: improve warn/error logs in the mux
These were broken by 47d1347, resulting in error messages like:

    materialized::mux: error handling connection: in pgwire server

with no information about the actual error. This patch fixes the issue
by ensuring that the underlying error is printed along with all of its
causes.

It also makes adding the name of the handler to the error message the
responsibility of the mux, rather than the handler, so that handlers
don't forget to add the appropriate context.

Fix #4441.