sql: auto-derive descriptions for SHOW queries
This avoids duplicating the logic for the various modifiers that change
the set of columns that are available, like "FULL" and "MATERIALIZED".

It also makes the names of the columns match the names that they have in
the query. This might be a bit disruptive, but it makes things much more
consistent. The changes are fully described in the release notes.