sql: derive an object's type from its ID
This follows the principles of normalization: don't include data that
can be derived from other fields. It also avoids a lot of joining in the
SHOW code.

Also, lowercase the types ("system" and "user" rather than "SYSTEM" and
"USER") so that it doesn't feel like we're screaming at our users.