coord: use default compaction for log indexes
mz_catalog log indexes have had a compaction time of 1s, instead of
the default logical-compaction-window. The removed comment documented
a general lack of knowledge about why this was, unsure if there was a
specific reason that these log indexes needed a shorter time, or if it
was that way because this number had just been diligently copied down
through many changes without a reason.

Investigation revealed the following commit:
https://github.com/MaterializeInc/materialize/commit/1cd5749e64ef5a146b1a9a557513e889c2d016ad.
This commit included comments "Constant chosen arbitrarily; choose better"
and "1000 ms compaction lag. Change!", describing that the 1s duration
is not meaningful.

We have also observed is #3983 that mz_catalog will often return a
timestamp error, which we suspected was contributed to by this short
compaction window. That problem was worked around (perhaps not quite
fixed) by #4467. But that workaround results in sometimes choosing a
future time and waiting, resulting in slow mz_catalog queries. Using
the default compaction window for mz_catalog will hopefully reduce the
amount or frequency of needing to wait.