dataflow: optimize reduce memory usage. (#4523)
Touches #4325

Reduce operations for hierarchical, inaccumulable reductions (think
non-monotonic mins and maxs) go through a staged reduction where initially,
very fine subsets of the group being reduced are handled independently, and slowly
those reductions are merged together to yield the final reduction.

In a lot of the initial stages, the number of buckets is so large that the reduction
ends up being a no-op, because each bucket only has one element. The number of
buckets has to be large because we might have to reduce a enormous data set,
and we can't change the structure of the "reduction tree" dynamically.

These changes make it so that we don't arrange both the input and the output at each stage
of the reduction. Instead, we arrange the input, and the elements not in the output. To give
a little bit of color with an analogy, our reduction is like a winner take all tournament, except
not nearly enough players enter most of the time. We end up spending a
bunch of memory recording things like:

"Player X vs. no-one -- Player X advances by default"

and we end up doing this so many times that we keep multiple duplicate
copies of the original set of players.

This change changes the recording to emit something analogous to

" Player X vs no-one -- no-one eliminated" and save memory on every empty set of
"eliminations". Astute sports fans will note that in a winner take all
tournament with N players, N - 1 players have to lose exactly 1 time, and the intent
of these changes is to arrange the entire dataset in the output of the reduce operator
exactly one time (as opposed to currently where the dataset is captured ~16x in the
reduce operator outputs).

This should reduce reduce memory usage by approximately 1/2 in the best case.