Convert the docker-compose version check to a warning, support 1.24
We don't know exactly what versions of docker-compose we test with. 1.24 is the version
that we test with in cloud, so we know it is supported. If folks observe that we work
with a lower version then we can make this number lower.

That said, the purpose of this check is to provide a bit of a warning in case users see
surprising behavior, hard-stopping them doesn't seem as useful.