Support debezium deduplication over a window of original event times
With this, users will be able to bound the amount of time that their materialized
clusters spend with linearly growing memory to track possible duplication.

This solution doesn't implement the full "arbitrary ranges" solution proposed in my
comment in #4712, since as far as we know debezium should not get into this
possibly-duplicating state without very specific and extremely unlikely scenarios.

Additionally, we don't have access to composite types (specifically lists) inside our
with options parsing code, so making this a more generic feature is both not obviously
ever going to need to grow more flexibility and making it look reasonable would require a
larger refactor of our sql-parsing code.

Resolves #4172