Change warn -> debug for double-deletion of metrics
These warn logs caused us to create MaterializeInc/rust-prometheus#12 , which I'll fix soon,
but as it is these were only put in at warn level to ensure that we did know if that was
a potential bug. Now that we know that it *is* a bug the warn is spurious and is
unnecessarily concerning, because there's nothing anyone can do about it.