sql: optimistically type LiteralArray and LiteralList
The two CoercibleScalarExprs LiteralArray and LiteralList previously
did not return a type when "typed" (i.e it returned None). However,
for polymorphic typing to work, values we plan as LiteralArray and
LiteralList need to express that they are arrays or lists,
respectively.

Because arrays and lists have an interior element type, we now
attempt to type those, as well. However, if we cannot determine the
type (e.g. they're all LiteralStrings), we identify the element as
a string. This, fortunately, does not reduce the expressive power
of our type system because strings can be cast to any type, which
permits us to recover the array's or list's "true" type using a
cast at some later point in planning.