wip: pushdown mfp
This is a draft of pushing down MFPs, right now it only really handles
FlatMap (and is not used in the optimization process). Putting it up now
to get some early eyes on it because it's pretty hard-to-follow at the
moment. I'm going to keep trying to make it simpler but this is the best
I've arrived at so far.