coord: unify bootstrap/steady-state catalog reporting
The various catalog tables (e.g., mz_databases) need to be updated
whenever the catalog state changes. This means once at startup, to make
the tables reflect the initial state of the catalog, and again on every
catalog operation.

Previously, there were two separate code paths for the startup and
steady-state updates. This patch unifies those code paths. The key
insight is that on startup, we can synthesize a list of events that
describes how to take a catalog from empty to its current on-disk state;
events that use exactly the same format as the events that are generated
for a steady-state catalog update.