Don't reuse a RowArena for MapFilterProject
In #4463 we introduced MapFilterProjects on top of Gets to reduce the
amount of data that we had to pack and unpack across operator
boundaries. When doing this, I made one RowArena that was used across
each invocation of MFP, which resulted in memory buildup over time. This
change changes this to occur per-row so that we can throw away
intermediate data.