expr: add regexp_match function
Provides for basic string manipulation capabilities. Matches the
PostgreSQL function of the same name, except that the regexps we support
are a bit different.

The name "regexp_match" conflicted badly with the internal naming of the
"~" operator, which was "MatchRegexp". Those functions have been renamed
"IsRegexpMatch", to more clearly indicate that they return a bool.

Fix #4545.