sqllogictest: fail during rewrite on unexpected error
If there's any kind of unexpected output we should fail during
rewrite. Otherwise some kind of "query I ..." statement that failed to
execute (i.e., should have been a "query error ...") will be silently
ignored. This is different from what we want, which is that the statement
succeeded, but the output is different.