coord: only update system tables for fully created sinks
Our setup process for sinks more complex than for other objects. First, we "reserve"
an id in the catalog, then we set up external state (for example we create a new Kafka topic
and register a schema in the schema registry) and then we delete the provisional object in the
catalog, and reinsert a new one.

Previously, we were not properly cleaning up state in system tables when we created a reserved catalog
object for sinks, but the external setup failed. This commit changes the behavior so that we only modify
system tables when we are sure that the sink will exist (e.g. after setting up external state).