sql-parser: rewrite lexer
The SQL lexer was some of our oldest and least idiomatic Rust code. It
was imported from the sql-parser project long ago, where it needed to
support lexing more than just the PostgreSQL dialect, and never got
fully simplified.

This commit is a nearly complete rewrite of the lexer, to make it both
more idiomatic Rust and more in line with PostgreSQL's lexical
structure. Notable changes:

  * Support for concatenating newline-separated string literals.
  * Support for dollar-quoted string literals.
  * Removal of whitespace tokens, which greatly simplifies the parser.
  * Support for all PostgreSQL operators, including user-definable
    operators. This also means that no changes to the parser will be
    required to add support for new PostgreSQL operators in the planner.

The larger motivation for this change is to support PostgreSQL's
qualified operator syntax, like OPERATOR(pg_catalog.~).