sql: support DISTINCT ON (exprs...)
`SELECT DISTINCT ON (exprs...)` is a PostgreSQL extension to `SELECT
DISTINCT` that allows the user to specify exactly which columns they
want to participate in the distinct-ing.

Our docs incorrectly claimed that we supported this feature. Though this
was technically a docs bug, we've had at least one user ask for this
feature, so seemed more userful to make the code match the docs here. So
there are no docs changes in this patch, because the docs for this
feature already exist.

Tests are nearly entirely provided by way of a CockroachDB SLT file that
was incorrectly removed in 30955d54c35. This patch restores that test
file and makes some minor modifications to the expected error messages
so that it passes with Materialize.

The initial version of this patch broke `SELECT DISTINCT ... ORDER BY`
that was not caught by the existing SLT tests, so the patch also adds
some ORDER BY-related tests.

Fix #4663.