coord: reorganize the session module
Reorganize the session module to be less fractured, so it is a bit
easier to follow. Now that all the vars are in their own file, I think
it is clearer to have everything else in the top level module. This also
avoids "module inception", which is a clippy lint for a reason: it's
very confusing to have a parent module with the identical name as one of
its children.