coord: always remember to replace mz_logical_timestamp
Optimizing an expression that calls mz_logical_timestamp will panic.
Previously, the coordinator was responsible for replacing calls to
mz_logical_timestamp, except for in contexts where the SQL planner had
promised not to permit calls to mz_logical_timestamp. This was hard to
keep track of, and several code paths miscommunicated on their
responsibilities.

Rework so that the SQL planner is always willing to plan a call to
mz_logical_timestamp, and the coordinator has the sole responsibility of
either replacing that call with the actual timestamp, or rejecting the
query.