avro dedupe: Add a deduplication_pad_end
This allows us to have a period after the end of deduplication where we do not
trust the deduplication information, but we keep it around and keep adding to
it in case we go back in time again.

This seems like it may be strictly worse than just extending the end of the
deduplication window to pad end, since especially at the end of the dedupe
window we should trust the full deduplication logic more than the "debezium
never time travels" invariant.