materialized: only notify about new versions
This ensures that users who somehow skew with the version cluster (e.g.,
due to our error) don't get erroneous notifications. It also ensures
that dev builds don't print out messages about *older* releases.

For good measure, we never print a warn! in a dev build, only an info!,
on the assumption that users running dev are more likely to be running a
custom build and don't want to deal with ugprade spam.