sql: support timestamp done messages in TAIL
Add a `WITH (TIMESTAMPS)` option to TAIL that adds a `done` bool column
to the output. At the end of each set of batches send a new row with
the timestamp, a true `done`, and NULL for all columns to indicate
the timestamp is done and no other rows will appear at or before that
timestamp. Otherwise `done` is false.

This works because each batch represents a disjoint and increasing bound
of time. Thus we only need to send a done message for the last batch of
a group of batches.

Closes #4539