sql: support timestamp progress messages in TAIL
Add a `WITH (PROGRESS)` option to TAIL that adds a `progressed` bool
column to the output. At the end of each set of batches send a new
row with the timestamp, a true `progressed`, and NULL for all columns
to indicate that no rows will appear before that timestamp. Otherwise
`progressed` is false.

This works because each batch represents a disjoint and increasing bound
of time. Thus we only need to send a progress message for the last batch
of a group of batches.

Closes #4539