pgwire: prepare for more error response fields
The present design of the error reporting interface in the pgwire crate
makes it very convenient to report errors that use the default fields
(code and message), but impossible to report errors that use additional
fields like hint and detail. Forthcoming commits will add even more
additional fields, like the position in the query at which the error
occurred.

Rework the API so that reporting an error requires building up an
ErrorResponse manually, with whatever fields they like. But add builder
methods to ErrorResponse so cut down on the verbosity. This results in a
slight hit to ergonomics in exchange for a big increase in flexibility.

Also, match PostgreSQL in using exactly the same type to represent
errors and notices. No point in enforcing a separation that doesn't
exist.