sql-parser: unify LIMIT and FETCH
The FETCH clause is the SQL standard way of specifying LIMIT. Unify the
two clauses in the parser so that both are supported by the planner.
Previously we were silently ignoring the FETCH clauses. (Note that FETCH
... WITH TIES is still not supported.)

This commit drops support for parsing FETCH ... PERCENT, since not even
PostgreSQL supports that.