sql: Add IdGen to StatementContext
Supporting CTEs requires either unification or translation of IDs
generated during planning vs. those generated during decorrelation.
This is due to the fact that references to CTEs use LocalIds, the same
namespace which are generated during decorrelation. @benesch
:recommended unification, so that's what this PR does.

By hoisting IdGen into a StatementContext, we can use it to generate
local IDs for CTEs (done in subsequent commits), as well as pass
it along to the decorrelation process.

The IdGen is stored inside of a RefCell in the StatementContext,
mirroring the parameter type map, because it is not clearly possible to
thread a mutable reference to the StatementContext through the planner.