coord: decouple portals and prepared statements
Previously portals would refer to the prepared statement from which
they were created. This was used to fetch a portal's description. The
upcoming feature to support DECLARE cursors is a way to create portals
that are not attached to a prepared statement.

Move the statement and its description to the portal so that portals do
not have to have an associated prepared statement.