support map types internally
Adds support for representing maps inside of Materialize. Adds a map variant
to the pgrepr Type and Value enums, and hardcordes a new, non-stable OID for maps.
Adds the ability to parse string literals as maps in order to test the new
internal representation.

Note: nested maps are currently unimplemented, usage of non-nested maps requires
the user to start Materialize in experimental mode.