sql: support DECLARE, FETCH, CLOSE cursor statements
The purpose of this support is to allow users with drivers that force
buffering all results in memory (most drivers) to be able to use TAIL
as a normal rows-returning statement (i.e., not COPY). FETCH makes it
easy to ask for a limited number of rows at a time.

These SQL statements are portal control messages that have exact pgwire
equivalents. As such they can be implemented without involvement of the
coordinator, similar to how the other portal control messages are (Parse,
Bind, Execute). In practice this means that one_query and execute now
have early-exit points for cursor statements. In addition, we also need
to pass around possible source portal information for correct semantics.

FETCH here is not a full implementation with respect to what Postgres
supports. Their FETCH drains the target portal of N rows, caches them,
and returns them as requested to users. Users could use Execute pgwire
messages to retrieve partial results at a time. We expect no one to do
this and, in order to avoid implementation difficulty, do not support
that use case.

sqllogictests are not included yet because it would require implementing
cursors there. Instead sqllogictest will be changed in a later PR to
work over pgwire, at which point cursor tests will be added.