lint: make errors more visible and consistent
Make linter errors more visible by using the format

    lint: SEVERITY: program: message

where SEVERITY is either "error" or "fatal" and is printed in red.
Use this format consistently across all the lint and lint-like scripts,
like pycheck. Also make things quieter when successful, so the errors
are easier to spot.

In the course of thsi change, clean things up so that lint and pycheck
share more code, since they are very similar in nature.