sql: plan CTEs another way
Plan CTEs as explicit Lets/Gets so that those Lets/Gets can be carefully
lowered to dataflow Lets/Gets. The previous implementation inlined CTEs
at the site of every use.

The new implementation better matches the model of CTEs as relation
expressions that are computed once and reused--though because
Materialize does not yet support impure functions or statements with
side effects in CTEs, there is no behavioral difference between this
implementation and the last.

Co-authored-by: Nikhil Benesch <nikhil.benesch@gmail.com>