Use named volume to hold Prometheus data
When /var/lib/docker and the Git checkout live on different filesystems,
Prometheus will fail to startup because it cannot mmap files the files
mounted into its container from the git checkout's filesystem. Use a
named volume instead, as this will allow the mmap to succeed. This will
continue to maintain data across restarts for workflows that don't
explicitly remove volumes. For volumes that do remove volumes, then this
will result in a behavior change, as the Prometheus data will be removed
as well.

Since we are no longer dumping data into the prometheus/data directory,
we can also remove the .gitignore. This also has the side effect of
reducing the amount of extra data in the git checkout.

Fixes #4847