sql: add -> map operator (#4846)
sql: support -> operator for maps

Adds support for using the -> operator with maps. The -> operator
allows users to fetch values from a map, given a single string key
or an array of string keys. For example:

    SELECT '{a=>1, b=>2}'::map[text=>int] -> 'a'

Returns 1.

    SELECT '{a=>1, b=>2}'::map[text=>int] -> ARRAY['a']

Returns {1}.

Any keys not found in the map will return a NULL value.