Add caveat re: Docker to get-started instructions
Before running through the getting started guide, I had previously run through the "install" guide and used the Docker setup instruction. I then came to the getting started guide and assumed running Materialize in Docker was an adequate substitute for installing and running the binary directly. Turns out it wasn't -- I got a `No such file or directory (os error 2)` error when trying to `create source` from the wikirecent file. I then tried installing via `apt` and the guide worked, so it seems that using Docker was the problem (perhaps it's looking for the source file on the image's file system, instead of the file system from which `psql` is being run).

It might make sense to instead provide a way of making the getting started guide compatible with Docker, but this warning should help until that can be done.