interchange: add map to AvroFlatDecoder
Adds the ability for the AvroFlatDecoder to decode map types.
Because the pairs contained in the map need to be packed into the
row in key-sorted order, the current implementation must decode each map
value using the ValueDecoder. This route is expensive, and should
be fixed in the future if possible.