sqllogictest: use a regular postgres connection
The previous implementation required correctly implementing many parts of
the pgwire protocol (prepare, bind, execute) itself, including dataflow
and coordinator handling. This meant that some SQL features would have
to be re-implemented here in order to be tested (TAIL, cursors). The
initial reason that slt could not use pgwire are now gone, so it is now
reasonable to make this a normal pgwire connector and remove the need
to double implement some things.

We now format some data (times with microseconds, decimals and floats)
differently in some cases. This is not a change in the actual data
over the wire, just our driver's string display of it (i.e., it doesn't
actually matter).

Fixes #2583