expr: take control of traversal order during reduction
This commit introduces no semantic change, the new code performs the
identical computation as the previous one. However, this new structure
allows the reduction logic to recurse in non-default ways when
apropriate. For example we stop processing a Coalesce early if we find
an appropriate expression and skip reducing all other expressions after
that.

This refactor is also the basis for improving the error propagation
during reduction which is tackled in a separate commit.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>