expr: short-circuit coalesce subexpression reduction
When processing a COALESCE we only need to reduce the argument
sub-expressions until we find a non-null literal or encounter an error.
This commit special cases the COALESCE case of the variadic functions to
allow for that. This saves a bit of work for the optimiser but also sets
the stage for switching the Literal variant of ScalarExpr from holding a
Result<Row, EvalError> to holding a Row directly.

Furthermore, this commit:
   * Fixes #4969
   * Uses a linear scan to dedup the expressions instead of a HashSet

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>