expr: transform ScalarExpr::reduce into a fallible fn
Here we change the error path of ScalarExpr evaluation from being
through the ScalarExpr::Literal(Err(..))) variant to surfacing directly
into the Result of reduce, similar to how the optimize() family of
functions in the transform work.

The observation is that Literal(Err(_)) is only ever constructed in the
reduce function is manually propagated upwards, as if it had the ?
operator applied.

Switching reduce to fallible and using the actual ? operator allows
stopping processing of arguments earlier than previously, so it ought to
be a bit faster. Additionally we get all the benefits of using Result
like the compiler complaining if a reduce error is ignored.

Finally, because reduce was the only place that the Literal(Err(_)) was
constructed, after this commit it never happens. This sets the stage for
completely removing this variant and simplify all other places in the
codebase that work with ScalarExprs.

The previous error paths are left intact with unreachable!() statements
for the ease of the reviewer. All of these are removed in the next
commit that removes the Err Literal variant.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>