dataflow: improve "bad UTF-8" error message during regex decode
The old log message was hard to visually parse because it included the
contents of the bad line undelimited in the message:

    WARN dataflow::decode::regex: Line this line has invalid UTF-8 and will be dropped --> � <-- from source materialize.public.regex_source_primary_idx cannot be decoded as utf8. Discarding line.

This commit rewrites the log message to be a bit easier to visually
parse:

    WARN dataflow::decode::regex: dropping line with invalid UTF-8 (source: materialize.public.regex_source_named_cols_primary_idx, line number: 3, line prefix: "this line has invalid UTF-8 and will be dropped --> � <--")

The new format also includes the line number when possible.

For good measure, it also rewrites the code that computes the prefix of
the line to display in the log message. 1024 characters is too many, in
my opinion, considering that the line number is available now, so the
limit is decreased to 64. The code also applies the limit *after* the
UTF-8 decoding, rather than before; the old approach had the defect that
it could introduce a UTF-8 error (by chopping a multibyte character in
half at the 1024 character mark) that was unrelated to the actual UTF-8
error in the input. The new approach results in an extra allocation, but
that's ok on the error path.