parser: reserve time-like keywords as column aliases
Time-like keywords, e.g. YEAR, are not permitted to be used as
aliases following an expression, unless prefixed with AS. This
restriction ensures that users do not confuse

```
SELECT pg_catalog.interval '1y 1m 1d' MONTH;
```

as equivalent to:

```
SELECT INTERVAL '1y 1m 1d' MONTH;
```

Without this change, the former is actually parsed such that "month" is
a column alias, as in

```
SELECT INTERVAL '1y 1m 1d' AS month
```

which is likely not what the user intended. This patch makes it so that
`SELECT pg_catalog.interval '1y 1m 1d' MONTH` produces an error.

Facilitating this change required tweaking the parser's various reserved
keyword sets. Any keyword that appears at the top level of a select
(FROM, WHERE, HAVING, etc.) is now considered fully reserved. This
ensures that `SELECT WHERE` produces a syntax error rather than parsing
as `SELECT "where"`, for similar reasons as above.

Co-authored-by: Nikhil Benesch <nikhil.benesch@gmail.com>