doc,test: clearly document/test support for krb5 sinks
It turns out we do support these, but it wasn't clearly documented. Make
it a bit clearer in the docs and add proper integration tests for
krb5/sasl_plain sinks.

Also make it a bit clearer in the code by prefixing calls to the
generically named "extract_config" with the module name ("kafka_util"),
as we do elsewhere in the SQL crate and as Rust convention dictates.

Fix #4956.