mzcompose: learn a new --mz-find option
This works similarly to the old `--find` option to mzconduct. The idea
is that a composition is named after the directory containing
mzcompose.yml, and that these names must be unique across the entire
repository. This gives the infrastructure repository more stable
identifiers for load tests, so that we can restructure the repository
without breaking the infrastructure repository.