demo: remove kafka-transactions demo
This was never quite a demo, but rather a handy tool for generating
transactional data in Kafka. It hasn't been used in quite some time, and
rust-rdkafka is about to learn to generate data transactionally. So
remove it, per the demos discussion we had over Zoom a while back.

If we ever do need this code, it will always be available in Git.