Use MZ_CHBENCH_SNAPSHOT instead of MZ_TOPIC_SNAPSHOTS
In both cases, this is a single directory with a bunch of files,
including and not limited to topic snapshots.

Also, treat the snapshot directory and snapshot id variables the same,
no matter which script is using them. This means that we need to
specify the snapshot ID when running snapshot bench.

Rename data-directory to snapshot-directory.