Include source_id in Kafka reader balancing
If a user creates a bunch of single-topic partitions, then worker-0
would be responsible for reading from all partitions. Include the source
id in the calculation for assigning sources so that single-topic sources
are less likely to have to same worker assignment.

We probably want to do some additional work for multiple-partition
topics where the number of partitions is less than the number of
workers. One can imagine a scenario where the current log still
results in unbalanced workers:

    Worker-0    Worker-1    Worker-2    Worker-3
    Src-A, 0    Src-A, 1
                Src-B, 0    Src-B, 1

Fixes #5088