testdrive: isolate esoteric tests
Move the tests that require esoteric infrastructure or configuration to
run into their own directory. This makes it so that using the glob
`test/testdrive/*.td`, like most developers do locally, picks up the set
of tests that are "easy" to run. CI, however, will automatically run the
tests in the esoteric directory.

The old "ci-only" directory actually contains one test that is not run
at all, so this commit also renames that directory to "disabled."

Fix #5104.