sqllogictest: optionalize timestamp output
The timestamp-prefixing of output added in c881378 makes the output
harder to read on a standard-width terminal. Move that behavior behind a
flag, --timestamps, so it is availble when necessary but not on by
default.

Also rework the timestamp-prefixing implementation so that it applies to
multi-line output, e.g., errors or SQL statements that span multiple
lines.