file-sources: Only create a file reader if the source is active
Before this we were creating one file reader per materialized worker, which
would read the entire file into memory, put it into a queue that was not
getting read from.

The queues were bounded to 10,000 records, so this was unlikely to cause an OOM
but it is a bit unfortunate.