doc/user: refactor the select statement grammar to be simpler
The railroad diagram generating website indicated that we didn't need to retype the inner
body of the select statement for select all, distinct and distinct on. I further reorganized
how the definition was displayed to make it easier to read. I regenerated the diagrams to
make sure this change didn't cause any changes to the generated diagrams.