sql: split plan::statement module into bite-sized chunks
This module was one of the largest in the sql crate and very hard to
follow as a result. Group the functions into thematic modules to make it
more manageable.

This commit is pure code movement, with the exception of some tweaks to
where the portal code handling lives so that the statement module
doesn't expose its innards to the pgwire crate.