pgwire: permit hangups in startup sequence
Clients sometimes hang up during the pgwire startup sequence, e.g.,
because they received an unacceptable response to an SslRequest. Don't
log an "early eof" error message in this case, since this is considered
a graceful termination in the pgwire protocol.

pgx has this hang up behavior in its default configuration, as does psql
in its "sslmode=require" configuration.

This commit also adds traces to the pgwire startup flow to mirror the
traces in the main pgwire flow.

Fix #5246.