sql: add a type variable to the AST
This change is the first step in teasing apart semantic analysis and the
construction of our IR. Currently these two things are interleaved,
which is fine, but makes it difficult if you want to leverage the
analysis we do without subsequently constructing a plan.

The eventual goal of this work will be to change this pipeline:

```
AST -> IR
```
to this one:
```
Raw AST -> Augmented AST -> IR
```

Where the augmented AST already has things like types checked and names
resolved.

With the design goal that it should be possible to go back to the raw
AST, or at least a stringified form, from the augmented AST, to allow
for things like safe renaming of identifiers. A secondary goal that I'm
not entirely sure is possible yet is for the augmented AST to contain
enough information such that the second step in the pipeline can be
performed without access to the catalog.