Improve error message for missing mzbuild services
When adding new services, it's common to forget to `git add` the
directory that contains the mzbuild.yml file. Before this change, that
would result in an error message about `sqllogictest/sqlite` being a
directory. After this change, the user should be made aware that the
problem is between their mzcompose.yml, mzbuild.yml and git.

The proper fix for this is to run `git add` on the mzbuild.yml file.

Fixes #5020