*: use structopt to parse command-line args
Our CLI tools variously used the getopts and structopt crates to parse
arguments. Having two separate command-line parsers floating around is
suboptimal, as it's two APIs that folks have to learn, and two slightly
different interfaces to learn.

While getopts is more flexible, it also introduces quite a bit more
boilerplate. Everyone besides me seems to like structopt quite a bit
more.

So this commit replaces all usage of getopts across the codebase with
structopt, with the exception of the materialized crate, which will be
switched over in a future commit.