coord: limit which statements can exec in transactions
Background: the postgres spec says that everything happens in a
transaction. For us, there are three kinds:

1. Explicit. Things in between `BEGIN` and `COMMIT`.
2. Implicit. If a user sends multiple statements in a single Query pgwire
message, these are wrapped in an implicit transaction. (With some caveats
for it is contains or is already in a `BEGIN`).
3. Single statements (we call this idle, postgres calls it something
else). This is a Query message with a single statement in it or a single
Execute message (neither of which are within a `BEGIN`).

Previously we allowed any kind of statement to happen anywhere. If it
had side effects, like `CREATE`, `DROP`, or `INSERT`, then they would
be visible immediately to all transactions before `COMMIT`, and even
if an error happened later on in the transaction or a `ROLLBACK` was
issued. That is, we completely ignored the transaction request.

Change this so that only some kinds of statements can happen in explicit
or implicit transactions. For now, any statement that causes any kind
of data change can only be run as a single (idle) statement, since we
don't have a way to preserve any transaction semantics around those. Any
statement that reads or deals with the session is allowed in explicit
or implict transactions.

Reads in transactions are still broken since they move time around,
but the change here is an improvement on what we had before.