testdrive: Refactor Error to accept a list of causes
There are some places where we will want to run several actions irrespective of
individual errors, but fail the overall task if any errors occur. This makes
that possible.

To reduce the impact of future changes along these lines, this creates a new
helper method and pushes harder to use methods instead of constructing Errors
directly.