ensure the source's time since downgrade is initially stale
Previously (a93a2e77d) this was changed to force creation of an initial
timestamp to avoid timestamping data with `1`. However, the "is it
stale" comparison is at millisecond granularity and computers are fast,
so setting the initial time to now() - frequency can still result in a
small portion of data being timestamped as `1`.

Ensure that the initial time since downgrade is strictly in the past by
subtracting an extra second.