periodically poll rdkafka's event queue
This fixes the kafka statistics logging that is enabled by
`statistics_interval_ms`.

rdkafka has an internal event queue used for communicating back to the
application (variously referred to as rep, the main queue, etc.)

In the refactor ~8 months ago that split kafka sources into a single
consumer with separate queues for partitions we neglected to continue
polling the main queue.