sql: fix CREATE OR REPLACE VIEW bug
A user found that views created by CREATE OR REPLACE VIEW statements
are currently able to depend on the view that is about to be dropped.
This was reported in #5272.

The fix is to check during planning if the new view is attempting to
depend on the view that is about to be dropped, and throw an error if
so.