coord: consult since frontier when choosing timestamps for sinks
When creating a sink based on an unmaterialized object, the coordinator
intends to choose the minimum possible `AS OF` time, so that the sink
includes all updates that could possibly be emitted from that object.
The determine_frontier previously used timestamp 0 for this purpose,
but this was rather defective, as timestamp 0 might be before the
"since" frontier of involved sources. The ship_dataflow method would
notice this defective situation and complain about it, then fast-forward
the time to a valid time.

This resulted in a lot of ugly error messages about a situation that is
not actually problematic. So this commit changes determine_frontier to
choose a valid timestamp, and teaches ship_dataflow to return an error
message to the user rather than logging an error message. Now the error
should never occur when using automatic timestamp selection, but be
surfaced directly to the client when explicitly specifying an invalid
`AS OF` time.

Note that this discussion applies to `TAIL` statements too, as they
internally create a sink.

Fix #5391.