sql: add name resolution pass
This change adds a name resolution pass prior to planning, where table
ids are stuffed in alongside the table names. This will hopefully pave
the way to storing view definitions as id references instead of the raw
names.

I'm not really happy with the testing setup here, I had to put the tests
for the name resolution in the coord crate since that's where we have
access to the testing catalog. Ideally I'd like to have a simpler "test
catalog" implementation that can be loaded in the planner, but I guess
this will do for now?