coord/dataflow: timestamp BYO sources exactly once.
Touches #2915.

This commit removes a lot of complexity from the coord and dataflow crates by
teaching the timestamper to keep track of sources on a per-source rather than
per source instance basis, and it timestamps BYO sources once.

As a downside, we now have to keep the full set of BYO timestamp bindings in
memory, but future work will allow us to compact those bindings.