Fixup zookeeper logs bind mount (#5477)
I'm not sure why this worked when I tested this locally, but the
cloud benchmarks exposed that Zookeeper wasn't coming up with the
previous identity. Turns out it's because there was an extra s at the
end of the directory name.

It doesn't help that the example from Confluent's docs looks like
(notice that they called it logs?):

    -v /vol2/zk-txn-logs:/var/lib/zookeeper/log