kgen: allow avro keys and avro key distributions
This commit adds support for Avro formatted keys to kgen. Additionally this commit
changes the `partitions` arg to `partitions-round-robin` to make it more obvious
that this argument exists only to round robin records regardless of the contents of
their keys and therefore it shouldn't be used (Kafka will default to distributing keys
uniformly across partitions) unless the key distribution is really skewed but the
data still needs to be distributed evenly.