gate sinked updates if an existing consistency topic is detected
If a sink is created and an existing consistency topic is detected, we
attempt to read the latest commited timestamp and gate publishing until
our own ts has passed it. This is used for exactly once - on restart an
mz instance should re-bootstrap using the same timestamp history as the
previous instance and only start publishing once it has 'caught up' to
where the previous instance left off.