coord: cleanup TAIL sinks at txn end
Previously TAILs were only dropped when a cancel message was received or
the connection was terminated by the client, and only the most recent TAIL
was dropped. This was correct before portals were correctly supported. Now
that they are and we have correct transaction/portal semantics, move the
TAIL (sink) dropping to be handled when a transaction is cleared. We now
also support dropping any number of sinks instead of only the most recent.

Temporarily remove the abitily of the cancel handler to cancel TAILs which
will require a new API to communicate between coordinator and session.

Informs #5242
Informs #5543