coord: structure many errors
Introduce structured errors in the coord crate. This makes it possible
to return errors to the client over pgwire that report a SQLSTATE code,
details, and a hint in addition to the text of the error message.

The design here was proposed in #5340.

Converting all errors to structured errors is a large task, so this
commit largely limits the conversion to a few obvious places.
Unconverted errors remain in the CoordError::Unstructured variant as an
anyhow::Error. We'll want to continue converting these errors as we go.

The impetus for this change is to enable returning a hint with the "no
such role" method proposed in #5555.