ore: introduce str quoting helper
Introduce a helper that wraps a string in quotes, escaping any contained
quotes. This is useful for error messages like

    view "blah" did not exist

where the user-supplied name could potentially contain quotes. The
helper will automatically quote those internal quotes, resulting in
e.g.:

    view "bad\"quote" did not exist

I took a quick pass at converting constructions like

    format!("view \"{}\" did not exist", view_name)

but there are many, many places where we ought to be wrapping
user-supplied data in quotes but are not. We'll just have to improve
those as we go.

Per @mjibson's request.