expr: make it impossible to get the nullability of literal wrong
Make MirScalarExpr::literal and friends automatically ascertain the
nullability of the literal by checking if the literal is null. This is
guaranteed to be correct, unlike the prior approach which required that
the caller get the nullability right.

We made the equivalent change to HirScalarExpr in the sql crate to good
effect.