coord: allow terminating unstarted SessionClients
Previously the coordinator's SessionClient required that you had
previously called SessionClient::startup in order to call
SessionClient::terminate. Relax that restriction, because the pgwire
crate does not abide by it.

Specifically, a protocol mismatch error could previously cause the
coordinator to panic; now the error is properly returned to the client.