s3 sources: Support scanning from multiple buckets
With the new factoring of code necessary to support SQS updates this is
actually easier to implement than the old logic of requiring a different source
for different buckets.

This also more closely and accurately models the way that S3 sources are forced
to work. Since SQS queues can subscribe to multiple SNS topics[1], so they can be
configured for multiple S3 buckets. And since they can be configured for
multiple S3 buckets we have the choice of either erroring if a user configures
the same queue for multiple topics or of handling it correctly.

Because of the way that I've architectured things, it's actually easier to do
the more ergonomic thing of allowing users to handle multiple buckets in one
source, and so this refactors the source syntax to explicitly allow it.

[1]: https://stackoverflow.com/questions/44254094