materialized: avoid infinite recursion footgun
Restructure some code to avoid the possibility of infinite recursion
where a compile error is expected. Specifically, the Mux's
handle_connection trait method delegates to an inherent method of the
same name; but if that method of the same name changes its name or type
(e.g. due to a refactoring), the Mux will silently start infinitely
recursing instead of producing a compilation error as you might expect.