doc: remove last references to dc.sh
dc.sh is the precursor to mzcompose, and it's long gone. This commit
updates the the peeker-config README to refer to mzcompose instead of
dc.sh. The other reference to dc.sh was in the Metabase demo walkthrough
in the developer docs. This doc was supplanted by an official version of
the demo in the user docs, so this commit just deletes the now outdated
walkthrough in the developer docs.

Fix #5614.