*: temporarily remove distributed mode
Very early on in its life, Materialize learned to support a distributed
mode (#228), powered by Timely's support for scaling computation across
multiple processes. Unfortunately we never found the time to
productionalize this mode, and as a result it is quite fragile. We do
not presently recommend that anyone use distributed mode in production.

The headline problem with distributed mode is that a single transient
network error will cause the entire process to crash. This is true of
the network communications managed by both Timely and Materialize. The
assumption that the network is reliable is baked in pretty deeply and is
therefore not easy to change on short notice.

Even though we discourage use of distributed mode, we carry around a lot
of complexity to support it:

  * The entire `comm` package exists only in service of distributed mode.
    It provides channels that work over TCP, so that the coordinator can
    communicate with workers in different processes.

  * `comm` channels are clunky to use, because you have to initiate a
    TCP connection before you can use them. This is a fallible
    operation, which means code interacting with these channels needs to
    be prepared to handle errors nearly everywhere... of course, none of
    it does, so we just have a lot of `unwraps` lying around.

  * `comm` channels route interthread communication through the kernel,
    which is... certainly not a performance enhancement, though it's
    unclear just how costly it is.

  * `comm` channels use up ports/file descriptors, which are not free
    resources. Users have reported crashes in Materialize due to file
    descriptor exhaustion.

  * Server startup is complicated because it has to handle the case of a
    process that is only hosting Timely workers, and not a coordinator
    or pgwire server or HTTP server.

  * Network configuration is complicated because a Materialize process
    needs to listen for incoming Timely worker traffic from other nodes,
    rather than just incoming user traffic. This increases the surface
    area that needs to be secured when exposing a materialized process
    to a public network.

This commit is a proposal to temporarily remove distributed mode and all
the complexity it entails. The plan is still very much to support a
distributed mode in the future, but to do so when we have the
engineering bandwidth to manage the complexity.

Specifically, this commit:

  * removes the `comm` package,

  * replaces all use of `comm` channels with channels from either
    `tokio` or `crossbeam_channel`, i.e., production-grade channels with
    proper semantics and performance,

  * eliminates the concept of distributed mode from server startup.

The immediate benefit is that we needn't worry about securing Timely's
internal communications in the cloud product. Other short-term benefits
include possible performance speedups (we should measure!) and
simplifications in the coordinator.

Of course, there is a long-term downside here, which is that we will
start to implicitly encode assumptions about the fact that Materialize
is a single-node system into the code. All those assumptions will have
to be rewound when we do pursue distributed mode in the future.