transform: don't propagate expression errors in ColumnKnowledge
The ColumnKnowledge transformation looks at every literal in the
expression. If any of those literals are errors, it returns those errors
to the client immediately. This behavior is arguably wrong, because
those errors might never be evaluated.

Rewire ColumnKnowledge to instead pack these errors into the
DatumKnowledge it builds, rather than returning them to the client.