coord: make operations in read transactions use the same timestamp
Previously all read statements in a transaction would choose their
own timestamp, which isn't very transactiony. Now, on first query
in a transaction, choose a timestamp at which all other reads in the
transaction will also take place. Do this by finding all tables and views
in the schemas referenced in the first query and choosing a timestamp
that is valid across all of them.

Restrict TAIL in transactions to a solitary statement. Having TAIL
co-exist with multiple other statements in a transaction doesn't have
an obvious expected behavior, and this is probably not a useful thing
anyway. However we do need TAIL to be executable in transactions because
it's recommended use is with DECLARE, which must be in a transaction.