expr: allow constant errors (#5669)
* transform: don't propagate expression errors in ColumnKnowledge

The ColumnKnowledge transformation looks at every literal in the
expression. If any of those literals are errors, it returns those errors
to the client immediately. This behavior is arguably wrong, because
those errors might never be evaluated.

Rewire ColumnKnowledge to instead pack these errors into the
DatumKnowledge it builds, rather than returning them to the client.

* expr: allow constant errors

Allow RelationExpr::Constant to represent constant errors. This makes it
possible for the optimizer to stuff any errors it generates during
constant folding back into the RelationExpr so that those errors occur
during execution of the query rather than during optimization.

This unblocks an issue uncovered in #5651.

* strip lifted maps from first constant row

* improve err display

* more bugfixes

* ensure all workers build the same dataflow

* fold constant errors too

* correct some pgtests again

Co-authored-by: Frank McSherry <fmcsherry@me.com>