coord: remove NoSessionExecute
Remove NoSessionExecute, to force all SQL query execution through the
same path. In particular, this forces the HTTP interface to construct a
single-use session rather than using the system session. The upshot is
that the memory visualization works again, as the session now has a
proper search path (whereas the system session does not).