coord: persistent tables
This commit demonstrates an approach to persist table data across restarts of
Materialize. The current approach is fairly-prototype level, but it supports
creating / inserting into / dropping tables. It persists binary data into
a append only file with 24 bytes of metadata:
4 byte header, 8 byte timestamp, 8 byte diff and 4 bytes of row length
followed by the row itself.

This commit has a number of issues, in no particular order:

* not a stable format across releases, because our Row serialization can changes
* bunch of prototype level error handling
* ignores system tables on restart
* hasn't been tested with temporary tables or really at all
* unclear how much of a perf regression it is
* no protection against on disk data corruption
* sort of duplicates a lot of caching code