coord: add CHAR and VARCHAR as builtin types
We do not currently include CHAR and VARCHAR as builtin types. Once
we start correctly tracking type dependencies in Materialize (#5577),
this will prevent users from creating objects that use CHAR and
VARCHAR because these types are not explicitly created on startup.
So, this change adds CHAR and VARCHAR to our builtin types.