ore: forward port Iterator::is_sorted and friends
Iterator::is_sorted is a forthcoming method to the standard library. It
is quite useful, so copy and paste the implementation from the standard
library into `ore` with an `mz_` prefix so we can use it before it
stabilizes.

Once the feature stabilizes, we can remove the implementation from ore.