sql: don't mishandle reflexive equality conditions
MySQL allows you to include a column in the SELECT list that is not
present in the GROUP BY

    SELECT t1.a FROM t1 JOIN t2 ON t1.a = t2.a GROUP BY t2.a

using a rudimentary analysis to determine whether the columns are
declared equivalent via a JOIN. We support this by searching the ON
clause for "trivial" column equivalences. If we find any, we mark the
right column as "unnameable" and add its name as an alias for the left
column.

It turns out this doesn't work in the degenerate case of

    SELECT t1.a FROM t1 JOIN t2 ON t1.a = t1.a

that is, when a column is declared equal to itself in an ON clause. In
this case, when we mark the right column as unnameable, we of course
also mark the left column as unnameable, and thus that query falsely
claimed that "t1.a" does not exist.

This commit fixes the issue by changing find_trivial_column_equivalences
to only surface equivalences if the columns involved in the equality are
distinct.

Fix #5765.