materialized: disable TLS v1.3 in test_tls
This commit disables use of TLS v1.3 in test_tls. This will put a stop
to the flakes for once and all.

Briefly, the design of TLS v1.3 causes unstable error messages when a
client fails to present a certificate. See the comment within for
details. It's not clear whether we should fix `postgres`/`hyper`,
OpenSSL, or the TLS specification, so for now we just sidestep the issue
by using TLS v1.2.