sql: resolve names as part of normalization
This PR currently has a few issues: first, there's an error message I
throw away because we currently detect it by looking at what names we
have. Not sure how to recover that. Second, there's currently a failing
test in the TLS code that I don't know how to debug, but am going to
dive into a bit later.

I also do not have the most confidence that this doesn't break things,
so I need to go through and give it all another testing pass. Throwing
it up to get some eyes on it to see if there are any obvious big
problems I'm missing!