Handle Avro over Kafka without a schema registry
The magic bytes only make sense and are only included if the user is using a
schema registry, so we should only check for them in that code.

This does "automatic" detection of whether or not we should check for schema
registry bytes based on whether we are using a schema registry, which may not
actually be correct, depending on whether or not folks are using confluent
clients.