coord,sql: temporarily remove CREATE SINK ... AS OF
`CREATE SINK ... AS OF` does not have sensible semantics on restart.
Creating a sink on a real-time source `AS OF` some timestamp is a
guaranteed way to cause Materialize to fail to boot on restart.

This commit temporarily removes the feature. We'll want to restore it
once we sort out how to unify `AS OF` behavior (see #2862).

As a side effect it fixes #5423. Since we no longer let users specify
`AS OF` timestamps, we can defer selection of the timestamp until the
moment we ship the sink dataflow, which allows us to ensure we choose a
valid time.