sql: validate typmod for qualified varchars
Previously we wouldn't do this validation if the name of the type was
qualified. This hacky change fixes the immediate problem but I think
there's still a lingering issue where someone could make a namespaced
type of name varchar that was a string and be incorrectly beholden to
this behaviuor. This is a quick change to unblock #5816 which causes
names to be qualified in more contexts.