bin/check: disable more clippy lints
Categorically disable clippy's "style" and "complexity" lints. These two
categories have been the most significant source of friction in our
codebase.

The hope with these clippy lints is that they'd *reduce* the amount of
bikeshedding, because we'd all just accept clippy's advice even if we
had some quibbles with a lint or two. Empirically that hasn't been true,
and it's not worth the constant bikesheds.

We still have the capacity to enable the few lints in these categories
that are worthwhile.