s3-sqs: Wait for a notification when configuring S3 SQS
This commit investigates whether the flakiness in our S3 SQS tests is due to a
race in AWS, where the configuration for SQS notifications does not fully take
effect immediately.

To ensure that that is not the case, we don't allow testdrive to continue on
from notification configuration until it has observed at least one valid
"object has been created" message.