explicitly retrieve default partition/replication config from kafka
Kafka brokers have a notion of a default partition count and replication
factor. Newer versions of Kafka support setting -1 for these values in
client requests to pick up the default - older versions (pre 2.4) do not
and will instead throw an unsupported error.

This diff adds a code path that explicitly calls the Kafka API to
retrieve the default configurations.

To eliminate modality, this code path is used even for newer versions of
Kafka that support API calls setting -1.

There is extremely poor test coverage of this area. Notably, we do not
have test coverage for older versions of Kafka or Kafka with multiple
brokers (and thus do not have test coverage for partition counts or
replciation factors > 1).