Expose channel utilization metrics in mz_message_counts
Keyed (id, worker), this table holds information about the messages any channel
has received so far. The computation for sent/received message counts
needs to happen in a somewhat convoluted manner, as we have to join to
the correct (id, worker) pair on the channels stream (for received, we
get the receiving worker's ID, and for sent, we get the sending
worker's ID).

Also, add (and fix) some tests that ensure the table gets properly populated.