python: Update black, fix formatting
A 2020 version of black updated to change the way that it formats arg lists and
lists, such that a trailing comma means "always do multiline formatting".

A couple of the changes in here are black being stricter about docstrings, but
many of them are just getting rid of trailing commas so that it preserves our
existing formatting.