kafka prometheus: source_id needs to match our offset_max metric
In particular, we need to connect the the mz_partition_offset_ingested with
mz_kafka_partition_offset_max, and so the kind of ID (SourceId vs
SourceInstanceId) needs to be identical.

With this we split the SourceInstanceId into its component parts, allowing us
to do the work at query time, at the small cost of no reduction in cardinality.

We can't just remove the source_instance part, because when if the same
unmaterialized view is referenced by multiple downstream materialized views,
each reader will get its own instance ID and offset, clobbering all others.