Write peeker config file during ingest benchmark
This unifies setup-ingest-benchmark and generate-snapshot as a single
workflow called prepare-snapshot. The peeker config file is expected by
both workflows. It used to be the case that generate-snapshot would do a
lot more work, and that is why it was a distinct step from
setup-ingest-benchmark. But now that the snapshot is based on archiving
filesystems, these steps are equivalent.