materialized: use correct kafka address in test_no_block
In CI, this test needs to use a non-localhost address. Wire that up via
a KAFKA_ADDRS environment variable. This will fix the test_no_block
flakiness. (The reason for the flake is that `localhost:9092` has never
resulted in a successful Kafka connection, but we recently got better at
observing the failed connection in the short time window that we check
for Kafka errors during `CREATE SOURCE`.)