sql: multi-partitioned Kafka start_offsets
Allows users to indicate a different start offset for each of a
topic's partitions. This looks like:
    CREATE SOURCE ... WITH (start_offset=[1, 2, ...])

Note: this is a little quirky! If only one start offset is provided,
this offset will only be applied to the first partition, not all
partitions. All partitions without a specified offset will start
at 0. If too many offsets are accidentally supplied, new partitions
that come online will be incorrectly truncated.