Two fixups for chbench ingest benchmark (#5942)
* Remove extraneous single quotes

In general, these quotes are not required. Additionally, this
substitution is causing docker problems because it things that this is
the name of a docker volume:

    ${MZ_CHBENCH_SNAPSHOT:-'./tmp'}

(Docker sees the literal string "'./tmp'" instead of "./tmp").

* Write peeker config file during ingest benchmark

This unifies setup-ingest-benchmark and generate-snapshot as a single
workflow called prepare-snapshot. The peeker config file is expected by
both workflows. It used to be the case that generate-snapshot would do a
lot more work, and that is why it was a distinct step from
setup-ingest-benchmark. But now that the snapshot is based on archiving
filesystems, these steps are equivalent.