Change negative multiplicity error message to be human-readable (#5977)
The reason that negative multiplicity errors are a problem is easy to explain
and understand, but the term "negative multiplicity" is confusing. Instead,
just describe the problem that exists and the most likely reason it happened.

This does not change the error message for any of the asserts around this
subject users should never see them and negative multiplicity is more
technically correct in their cases.