Fixup lint error in mzbench
I'm not sure how this crept into main -- it appears to have been a bad
merge squash by GitHub. Git annotate shows that this flag was added as
part of #5912, the PR was set to automerge and the testing status is
green for the last commit on that PR. It's unfortunate that this crept
into the codebase -- clean it up.