Raise Dataflow errors for simple protobuf and avro (#5995)
* Raise Dataflow errors for simple protobuf and avro

We don't have any tests, at the moment, that are failing due to this
change because there are no tests for this code path. Still need to
update our tests for this code path.

* Test error streams for corrupted protobuf messages

Add an option to testdrive's ingest-topic command that enables
corrupting messages in an encoding aware fashion (currrently just XORs
the data with itself, which corrupts most things). The implementation of
each corruption is left to the transcoder.

* Add testdrive cases for Avro decode errors