dbt-materialize: fix catalog query (#6068)
In a very thorough bug report (#6063), it was pointed out that the
dbt-materialize adapter's catalog query performs a join between types
in the mz_columns and mz_types system tables that returns NULL for
aliased types. This faulty join causes columns to be without a type
in the generated dbt docs.

As suggested in the bug report, this change simply removes the join
and uses the type present in mz_columns, instead.