dataflow: remove can_close_timestamp
The logic for deciding whether or not a timestamp can be closed is now identical across
source implementations. Broadly:

A worker can close a timestamp t if all of the source partitions it is responsible for
can close t, which happens once we have read at an offset greater than the one
assigned as the final offset bound to t across all of the worker's partitions.