coord,sql: expose "session ID" via mz_session_id function
The session ID is used by telemetry to identify a particular instance of
a materialized process. Expose this ID to SQL via a new mz_session_id
function that mirrors mz_cluster_id.

The term "session" conflicts badly with the SQL concept of the same
name, so this function is tucked away in the "mz_internal" namespace for
now.