ore: simplify retry API
The retry API in the ore crate has grown organically over time,
resulting in IMO a fairly confusing, non-orthogonal API consisting of
a RetryBuilder object, a Retry object, and a retry_for helper method.

This commit simplifies the API, unifying RetryBuilder, Retry, and
retry_for into one object named Retry. This object has a default
constructor which picks sensible retry defaults for fallible network
operations, but exposes builder methods in the case that any of the
parameter methods need tweaking.

The configuration options are whittled down to:

  * initial_backoff
  * max_backoff
  * max_tries

Exposing only these three knobs provides a good deal of flexibility
while also making it clear that Retry does not attempt to limit how long
the actual fallible operation takes. The old API

    retry_for(Duration::from_secs(60), |_| /* op */)

implied that the function would time out after 60s, when in fact the
duration only controlled the maximum permitted cumulative sleep; the
time taken by the actual fallible operation was not taken into account.

As an added bonus, the new code comes with usage examples and tests.

I've ported all callers of the old API to the new API. In most cases
I've adjusted callers that specified a custom `retry_for` duration to
just use the new default parameters, since the various hardcoded values
were not (AFAIK) chosen with a great deal of thought.  But to be safe,
anywhere where the default would result in retrying for *less* time,
I've increased the `max_tries` parameter so that the new code will retry
for at least as long as the old code would have.

For additional context, see:
https://github.com/MaterializeInc/materialize/pull/5811#issuecomment-781737833