ore: simplify retry API
The retry API in the ore crate has grown organically over time,
resulting in IMO a fairly confusing, non-orthogonal API consisting of
a RetryBuilder object, a Retry object, and a retry_for helper method.

This commit simplifies the API, unifying RetryBuilder, Retry, and
retry_for into one object named Retry. This object has a default
constructor which picks sensible retry defaults for fallible network
operations, but exposes builder methods in the case that any of the
parameter methods need tweaking.

The configuration options are whittled down to:

  * initial_backoff
  * clamp_backoff
  * max_tries
  * max_duration

The most important change, in my opinion, regards calls like

    retry_for(Duration::from_secs(60), |_| /* op */)

in the old API. Previously that did not account for the amount of time
spent waiting for the actual operation to complete. Rather it would
retry until 60s of accumulated backoff. In the new API, the equivalent
construction

    Retry::default().max_duration(Duration::from_secs(60)).retry(|_| /* op */)

does take the amount of time that the op runs for into account.

As an added bonus, the new code comes with usage examples and tests.

For additional context, see:
https://github.com/MaterializeInc/materialize/pull/5811#issuecomment-781737833