RDKafka Logger WIP
This commit adds a logger that (perhaps poorly) duplicates the
Materialized Logger, with initialization, and plumbs it through to every
source. The intent is that the new logger will become Send + Sync safe.

Only the Kafka source cares about this new logger and it attempts to put
this logger into the GlueConsumerContext object that rdkafka uses as a
context for callbacks. Doing so causes the same compilation erorr that
would happen if we tried to use the MaterializeLogger in this place:

    `Rc<RefCell<(dyn for<'r, 's> FnMut(&'r Duration, &'s mut
    Vec<(Duration, usize, MaterializedEvent)>) + 'static)>>` cannot be
    shared between threads safely within `GlueConsumerContext`, the
    trait `std::marker::Sync` is not implemented for `Rc<RefCell<(dyn
    for<'r, 's> FnMut(&'r Duration, &'s mut Vec<(Duration, usize,
    MaterializedEvent)>) + 'static)>>` required because it appears
    within the type `timely::logging_core::Logger<MaterializedEvent,
    usize>` required because it appears within the type
    `GlueConsumerContext`

Next step: fix this compiler message.