Tracking kafka consumer statistics now works
Instead of trying to create a new Logger class that implements Send +
Sync, use crossbeam to send messages from the callback context into the
source info context. This will activate the source, triggering the
source to read messages from the callback and update the system log
tables. When a source is dropped, the statistics are removed.